% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileprocessing.R
\name{center_scale}
\alias{center_scale}
\title{center_scale}
\usage{
center_scale(x, scale = TRUE)
}
\arguments{
\item{x}{Input matrix. Make sure it is a matrix. Typically, row is samples, and column the features.}

\item{scale}{Logical, whether to scale the data or not. Default is \code{TRUE}.}
}
\value{
Outputs a list containing centered matrix.
}
\description{
data centering function with scale opiton
}
\details{
This function is needed to pre-process data when conducting plsda analysis.
         And the function can also be used for other purposes when needed.
         With the colSds function from matrixStats pacakge to calcualte column standard deviation,
         this function is faster than the native \code{scale} function when \code{scale = TRUE}.

         The scale here really means "z-score normalization": (x - mean(x)) / SD
}
\examples{
\dontrun{
centerX <- center_scale(data, scale = TRUE)
}
}
