% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioUtil.R
\name{convertToFactor}
\alias{convertToFactor}
\title{convertToFactor}
\usage{
convertToFactor(dfm, cols)
}
\arguments{
\item{dfm}{data.frame. Input data.frame.}

\item{cols}{str or vector of strings. Column names to convert.}
}
\value{
data.frame with columns converted into factors.
}
\description{
Convert data.frame features into factors.
}
\details{
The output data.frame has the same dim() as the input data.frame.
}
\examples{
\dontrun{
  convertToFactor(dfm = dfm, cols = c("gender", "regions"))
}
}
