% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioUtil.R
\name{.assign_group_folds_simple}
\alias{.assign_group_folds_simple}
\title{.assign_group_folds_simple}
\usage{
.assign_group_folds_simple(groups, group_mean_y, cross.k)
}
\description{
Internal helper 2 of 2: group-level fold assignment for nested CV
   Simple greedy regression group assignment: balance by unweighted group mean
}
