% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioUtil.R
\name{.assign_group_folds_stratified}
\alias{.assign_group_folds_stratified}
\title{.assign_group_folds_stratified}
\usage{
.assign_group_folds_stratified(sampleIds, y, cross.k, class_table)
}
\description{
Internal helper 1 of 2: group-level fold assignment for nested CV
   Stratified group assignment: for each class level, distribute that class's groups to folds
}
