% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_plsda.R
\name{dummy}
\alias{dummy}
\title{dummy}
\usage{
dummy(x, drop2nd = FALSE)
}
\arguments{
\item{x}{Input vector. Make sure it is a factor.}

\item{scale}{Logical, whether to scale the data or not. Default is \code{TRUE}.}
}
\value{
Outputs a matrix with dummified variables.
}
\description{
dummification of categorical variables
}
\details{
This function is needed to pre-process data when conducting plsda analysis. And the function can also be used for other purposes when needed.
}
\examples{
\dontrun{
y <- dummy(y)
}
}
