% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileprocessing.R
\name{fs_csv_generator}
\alias{fs_csv_generator}
\title{fs_csv_generator}
\usage{
fs_csv_generator(
  objTitle = "data",
  fltdata = NULL,
  rmControl = FALSE,
  anno = NULL,
  geneSymbol.var = "GeneSymbol",
  tgt = NULL,
  group.var = "Group",
  de.dfm = NULL,
  probeID.var = "ProbeName",
  fc.var = "logFC",
  p.var = "P.Value",
  adjP.var = "adj.P.Val",
  q.value = 0.05,
  fc.value = 1.5,
  fc.log.trans = TRUE
)
}
\arguments{
\item{objTitle}{Prefix for the output file name. Default is \code{"data"}.}

\item{fltdata}{Filtered expression data from microarray or RNAseq experiments.}

\item{rmControl}{Only functional for Agilent platform microarray datasets, to remove control probes. For other datasets, set to \code{FALSE}. Default is \code{FALSE}.}

\item{anno}{Gene annotation dataframe.}

\item{geneSymbol.var}{Name for the gene symbol variable from \code{anno}. Default is \code{"GeneSymbol"}.}

\item{tgt}{Sample annotation data frame.}

\item{group.var}{Name for the sample group variable from from \code{tgt}. Default is \code{"Group"}.}

\item{de.dfm}{Differential expression resutls dataframe.}

\item{probeID.var}{Name for the probe ID variable from \code{de.dfm}. Default is \code{"ProbeName"}.}

\item{fc.var}{Name for the fold change variable from \code{de.dfm}. Default is \code{"logFC"}.}

\item{p.var}{Name for the raw p value variable from \code{de.dfm}. Defaults is \code{"P.Value"}.}

\item{adjP.var}{Name for the adjusted value variable from \code{de.dmf}. Default is \code{"adj.P.Val"}.}

\item{q.value}{P value threshold for subsetting the filtered data. Default is \code{0.05}.}

\item{fc.value}{Fold change threshold for subsetting filtered data. Default is \code{1.5}.}

\item{fc.log.trans}{Whether the input fold change is log2 transformed. Default is \code{TRUE}.}
}
\value{
Outputs a \code{csv} file to working directory for RBioFS. If not \code{Elist} class from limma, Make sure that \code{fltdata} is a list object with \code{genes} (name) dataframe for gene annotation and \code{E} (name) as expression matrix/dataframe. For \code{tgt}, make sure to have sample name as rownames.
}
\description{
Create CSV file from microarray or RNAseq experiments ready for RBioFS
}
\details{
The function works best with the results from RBioArray.
}
\examples{
\dontrun{
fs_csv_generator(objTitle = "test", geneSymbol.var = "Gene.Symbol", fltdata = fltdata, anno = annot,
de.dfm = diana_medaka_DE$HTvC, tgt = Tgt, group.var = "Treatment")
}
}
