% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioUtil.R
\name{na_summary}
\alias{na_summary}
\title{na_summary}
\usage{
na_summary(data, by = c("row", "col"))
}
\arguments{
\item{data}{data.frame, matrix list of data.frames or matrices. Input data.}

\item{by}{str. Summarize by row or column. Default is \code{"row"}.}
}
\value{
Vector or list of vector of NA ratios.
}
\description{
NA ration summary.
}
\examples{
\dontrun{
  na_summary(data = data, by = "row")
}
}
