% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_plsda.R
\name{randomiz.test}
\alias{randomiz.test}
\title{randomiz.test}
\usage{
randomiz.test(residualsNew, residualsReference, nperm)
}
\description{
Randomization function from \code{pls} pacakge. For \code{\link{rbioClass_plsda_ncomp_select}} function.
}
