% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioUtil.R
\name{rbio_shap_svm_label_prob}
\alias{rbio_shap_svm_label_prob}
\title{rbio_shap_svm_label_prob}
\usage{
rbio_shap_svm_label_prob(object, newdata, col_idx = NULL, ...)
}
\arguments{
\item{object}{A SVM model object, in \code{rbiosvm} class.}

\item{newdata}{Input data to be classified. Make sure it is a \code{matrix} class and has the same variables as the model, i.e. same number of columns as the training data.}

\item{col_idx}{Probability column (i.e. label or outcome) index or name for which the probability vector is extracted.}

\item{...}{Additional parameters for the \code{e1071mc:::predict.svm} method.}
}
\value{
Vector or data frame of prediction probabilities.
}
\description{
Helper function to extract probability vectors from svm model predictions.
}
\details{
1. \code{newdata} should not include outcome column.
   2. \code{newdata} should be 0-1 scaled.
   3. \code{newdata} will be standardized using training data's col mean and sd, if the input \code{rbiosvm} model \code{object} contains standardization information:
      \code{object$center.scaledX}.This means the model was trainined using standardization.
}
