% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_plsda.R
\name{rbioClass_plsda}
\alias{rbioClass_plsda}
\title{rbioClass_plsda}
\usage{
rbioClass_plsda(x, y, ncomp = length(unique(y)) - 1, method = "simpls",
  scale = TRUE, validation = c("none", "CV", "LOO"), segments = 10,
  segments.type = "random", jackknife = TRUE, ..., verbose = TRUE)
}
\arguments{
\item{x}{Input data matrix (e.g., independent variables, predictors, features, X, etc). Make sure it is either a matrix or a dataframe.}

\item{y}{Input response variable (e.g.,dependent variables, Y etc). Make sure it is \code{factor} class or a character vector.}

\item{ncomp}{Number of components to be used for modelling. Default is \code{length(unique(y)) - 1}.}

\item{method}{PLS-DA modelling method. Four PLSR algorithms are available: the kernel algorithm ("kernelpls"), the wide kernel algorithm ("widekernelpls"),
SIMPLS ("simpls") and the classical orthogonal scores algorithm (also known as NIPALS) ("oscorespls"). Default is the popular \code{"simpls"}.}

\item{scale}{Logical, whether to scale the data or not. Default is \code{TRUE}.}

\item{validation}{Cross validation methods. Options are "none", "CV" (fold), "LOO" (leave-one-out). Default is \code{"CV"}.}

\item{segments}{Set only when \code{validation = "CV"}, the number of segement to be set. Default is \code{10}.}

\item{segments.type}{Method to set up the segments. Options are \code{"random", "consecutive", "interleaved"}. Default is \code{"random"}.}

\item{jackknife}{If to use jack-knife procedure. Default is \code{TRUE}.}

\item{...}{Additional arguments for \code{mvr} function from \code{pls} pacakge.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}
}
\value{
Returns a PLS-DA model object, with classes "mvr" and "rbiomvr".

       Additional items for \code{rbiomvr} object to \code{mvr} object from pls package:

       \code{centerX}: centered X data if applicable, with scale option.

       \code{dummy_y}: dummified y

       \code{inputX}: raw input predictor data.

       \code{inputY}: input group labels.

       \code{validation_method}

       \code{validation_segments_type}
}
\description{
PLS-DA modelling
}
\details{
The center and scale processes are handled by \code{\link{center_scale}} function.
         Therefore, the \code{center} and \code{scale} arguments \code{pls::mvr} are set to \code{FALSE}.

         The function always centers the data.

The data is always centered prior to modelling, i.e. \code{x - col.mean}. Thus no "center = TRUE/FALSE" option is provided.

         For \code{ncomp} value, the default (full model) compares feature number with \code{class number - 1}, instead of observation number.

         For sequencing data, the input x needs to be either tranformed with the function like \code{clr_ilr_transfo()} from \code{RBioArray} package,
         or normalized using methods like "TMM" or "RLE" implemented in \code{edgeR} pacakge.
}
\examples{
\dontrun{
rbioClass_plsda(x, y, ncomp = 20)
}
}
