% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_plsda.R
\name{rbioClass_plsda_ncomp_select}
\alias{rbioClass_plsda_ncomp_select}
\title{rbioClass_plsda_ncomp_select()}
\usage{
rbioClass_plsda_ncomp_select(
  object,
  ...,
  ncomp.selection.method = c("1err", "min", "randomization"),
  randomization.nperm = 999,
  randomization.alpha = 0.05,
  rmsepplot = TRUE,
  plot.rightsideY = TRUE,
  plot.optm.ncomp.line = TRUE,
  multi_plot.ncol = length(dimnames(object$coefficients)[[2]]),
  multi_plot.nrow = 1,
  multi_plot.legend.pos = c("bottom", "top", "left", "right"),
  plot.display.Title = TRUE,
  plot.SymbolSize = 2,
  plot.fontType = "sans",
  plot.xLabel = "Components",
  plot.xLabelSize = 10,
  plot.xTickLblSize = 10,
  plot.yLabel = "RMSEP",
  plot.yLabelSize = 10,
  plot.yTickLblSize = 10,
  plot.legendSize = 9,
  plot.Width = 170,
  plot.Height = 150,
  verbose = TRUE
)
}
\arguments{
\item{object}{A \code{rbiomvr} or \code{mvr} object. Make sure the object is generated with a \code{validation} section.}

\item{...}{Additional argument for \code{RMSEP} function from \code{pls} package.}

\item{ncomp.selection.method}{Optimal numbers of components selection method. Options are \code{"min"}, \code{"1err"}, and \code{"randomization"}. Default is \code{"1err"}.}

\item{randomization.nperm}{Set only when \code{ncomp.selection.method = "randomization"}, number of permutations. Default is \code{999}.}

\item{randomization.alpha}{Set only when \code{ncomp.selection.method = "randomization"}, alpha for the p values used during "randomization" selection. Default is \code{0.05}.}

\item{rmsepplot}{If to generate a RMSEP plot. Default is \code{TRUE}.}

\item{plot.rightsideY}{If to show the right side y-axis. Default is \code{FALSE}. Note: doesn't seem to be necessasry as PLS-DA always has at least two y classes.}

\item{plot.optm.ncomp.line}{If to display the vertical line indicting the optimal number of components. Default is \code{TRUE}.}

\item{multi_plot.ncol}{Number of columns on one figure page. Default is the number of responding classes, i.e. y.}

\item{multi_plot.nrow}{Number of rows on one figure page. Default is \code{1}.}

\item{multi_plot.legend.pos}{The legend position. Only effective when multi-plot is generated. Options are \code{"bottom"}, \code{"top"}, \code{"left"} and \code{"right"}. Default is \code{"bottom"}.}

\item{plot.display.Title}{If to show the name of the y class. Default is \code{TRUE}.}

\item{plot.SymbolSize}{Symbol size. Default is \code{2}.}

\item{plot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is avaiable on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{plot.xLabel}{X-axis label. Type with quotation marks. Could be NULL. Default is \code{"Components"}.}

\item{plot.xLabelSize}{X-axis label size. Default is \code{10}.}

\item{plot.xTickLblSize}{X-axis tick label size. Default is \code{10}.}

\item{plot.yLabel}{Y-axis label. Type with quotation marks. Could be NULL. Default is \code{"RMSEP"}.}

\item{plot.yLabelSize}{Y-axis label size. Default is \code{10}.}

\item{plot.yTickLblSize}{Y-axis tick label size. Default is \code{10}.}

\item{plot.legendSize}{Legend size. Default is \code{9}.}

\item{plot.Width}{Plot width. Default is \code{170}.}

\item{plot.Height}{Plot height. Default is \code{150}.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}
}
\value{
Prints the selected number of components for each y class. Returns RMSEP values for each y class to the environment as a \code{rbiomvr_ncomp_select} object,
        as well as a pdf file for the RMSEP plot if \code{rmsepplot = TRUE}.
}
\description{
Optimal number of components selection for PLS-DA model, with RMSEP plot funcitonality. Selection methods are modified based on \code{selectNcomp()} from \code{pls} pacakge.
}
\details{
The RMSEP figure shows both CV estimates and adjusted CV estimates, which is CV estimiates corrected for bias.
         Three methods are used for components number selection: \code{"min"} simply chooses the number of components to
         reach te minimum RMSEP; \code{"1err"} chooses the number of components when its RMSEP first reaches minimum as well as within one standard error;
         For "randomization", see the help file for \code{selectNcomp()} function from  \code{pls} pacakge.
}
\examples{
\dontrun{
rbioClass_plsda_ncomp_select(new_model,  multi_plot.ncol = 2, multi_plot.nrow = 2, plot.optm.ncomp.line = T,
         ncomp.selection.method = "randomization", randomization.nperm = 999, randomization.alpha = 0.05)
}
}
