% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_plsda.R
\name{rbioClass_plsda_perm}
\alias{rbioClass_plsda_perm}
\title{rbioClass_plsda_perm()}
\usage{
rbioClass_plsda_perm(object, ncomp = object$ncomp, adjCV = FALSE,
  perm.plot = TRUE, ..., perm.method = "by_y", nperm = 999,
  parallelComputing = TRUE, n_cores = parallel::detectCores() - 1,
  clusterType = "PSOCK", verbose = TRUE)
}
\arguments{
\item{object}{A \code{rbiomvr} object. Make sure the object is generated with a \code{validation} section.}

\item{ncomp}{Model complexity, i.e. number of components to model. Default is \code{object$ncomp}, i.e. maximum complexity.}

\item{adjCV}{If to use adjusted CV, i.e. CV adjusted for unbalanced data. Default is \code{FALSE}.}

\item{perm.plot}{Wether to produce a plot or not. Default is \code{TRUE}.}

\item{...}{Additional argument for \code{\link{rbioUtil_perm_plot}}.}

\item{perm.method}{Permutation method. Options are \code{"by_y"} and \code{"by_feature_per_y"}. Default is \code{"by_y"}. See details below.}

\item{nperm}{Number of permutations to run. Default is \code{999}.}

\item{parallelComputing}{Wether to use parallel computing or not. Default is \code{TRUE}.}

\item{n_cores}{Only set when \code{parallelComputing = TRUE}, the number of CPU cores to use. Default is \code{detectCores() - 1}, or the total number cores minus one.}

\item{clusterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}
}
\value{
The function returns \code{CSV} files for all intermediate permuatation RMSEP values as well as the p-value resutls.

The results are also exported to the environment as a \code{rbiomvr_perm} object with the following items:

\code{perm.method} Permutation method.

\code{nperm} The number of permutation runs.

\code{perm.stats} The stats metric used for the permutation test.

\code{adjCV} If the adjusted cross-validation stats is used.

\code{p.value.summary} P values for the permutation test.

\code{poerm.results} The intermediate permutation results, i.e. stats for each permutation test run in a data.frame. \code{nperm = 0} is the original stats.

A scatter plot is also generaeted when \code{perm.plot = TRUE}.
}
\description{
Permutation test for PLS-DA models.
}
\details{
The function uses RMSEP as the stats for comparing original model with permutatsions.

         Usually, we use the optimized PLS-DA model for \code{object}, which can be obtained from functions \code{\link{rbioClass_plsda}} and \code{\link{rbioClass_plsda_ncomp_select}}.

         Data for permutation are object$centerX$centerX, meaning centered X are used if applicable.

         Permutation methods are according to:

            Ojala M, Garriga GC. 2010. Permutation test for studying classifier performance.
            J Mach Learn Res. 11: 1833 - 63.

         For \code{perm.method = "by_y"}, labels (i.e. y) are permuated. A non-signifianct model (permutation p value > alpha, i.e. 0.05) in this case means the data is independent from the groups.

         For \code{perm.method = "by_feature_per_by"}, X is first subset by label (i.e.y) before permutating data for each feature.
         Since the permutation is done for the features WITHIN the group,
         the test actually evaluates if the model will produce significantly different performannce from the permutation models with the original "betweeen-features" relation (if any) disturbed.
         Therefore, A non-significant result (permutation p value > alpha, i.e. 0.05) means either the features are independent, or the model doesn't consider correlation between the features.
}
\examples{
\dontrun{
rbioClass_plsda_perm(object = new_model_optm, perm.method = "by_y", nperm = 999, adjCV = TRUE, parallelComputing = TRUE)
}
}
