% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_plsda.R
\name{rbioClass_plsda_scoreplot}
\alias{rbioClass_plsda_scoreplot}
\title{rbioClass_plsda_scoreplot}
\usage{
rbioClass_plsda_scoreplot(object, y = NULL, comps = c(1, 2),
  plot.rightsideY = FALSE, plot.Title = NULL,
  plot.sampleLabel.type = "none", plot.sampleLabel.vector = NULL,
  plot.sampleLabelSize = 2, plot.sampleLabel.padding = 0.5,
  plot.SymbolSize = 2, plot.ellipse = FALSE, plot.ellipse_conf = 0.95,
  plot.mtx.densityplot = FALSE, plot.mtx.stripLblSize = 10,
  plot.xAngle = 0, plot.xhAlign = 0.5, plot.xvAlign = 0.5,
  plot.fontType = "sans", plot.xLabelSize = 10, plot.xTickLblSize = 10,
  plot.yLabelSize = 10, plot.yTickLblSize = 10, plot.legendSize = 9,
  plot.Width = 170, plot.Height = 150, verbose = TRUE)
}
\arguments{
\item{object}{A \code{rbiomvr} or \code{mvr} object. Make sure the object is generated with a \code{validation} section.}

\item{y}{Set when object class is \code{mvr}. Input response variable (e.g.,dependent variables, Y etc). Make sure it is \code{factor} class.}

\item{comps}{Integer vector. Components to plot. The index of the components are intergers. The vector length should be between 1 and the total number of components, inclusive. Can be Default is \code{c(1, 2)}.}

\item{plot.rightsideY}{If to show the right side y-axis. Only applicble when the length of \code{comps} is less than 2, inclusive. Default is \code{FALSE}.}

\item{plot.Title}{Scoreplot title. Default is \code{NULL}.}

\item{plot.sampleLabel.type}{If to show the sample labels on the graph. Options are \code{"none"}, \code{"direct"} and \code{"indirect"}. Default is \code{"none"}.}

\item{plot.sampleLabel.vector}{Set only when \code{plot.sampleLabel.type} is not set to \code{"none"}, a character vector containing annotation (i.e. labels) for the samples. Default is \code{NULL}.}

\item{plot.sampleLabelSize}{Only set when \code{plot.sampleLabel.type} is not \code{"none"}. The size of the sample label. Default is \code{2}.}

\item{plot.sampleLabel.padding}{Set only when \code{plot.sampleLabel.type = "indirect"}, the padding between sample symbol and the label. Default is \code{0.5}.}

\item{plot.SymbolSize}{Symbol size. Default is \code{2}.}

\item{plot.ellipse}{If to draw ellipses. Default is \code{FALSE}.}

\item{plot.ellipse_conf}{The confidence value for the ellipses. Default is \code{0.95}.}

\item{plot.mtx.densityplot}{If to display a density plot on the diagonal for the correlation scoreplot matrix. Default is \code{FALSE}.}

\item{plot.mtx.stripLblSize}{The label font size for the correlation scoreplot matrix strips. Default is \code{10}.}

\item{plot.xAngle}{The rotation angle (degrees) of the x-axis marks. Default is \code{0} - horizontal.}

\item{plot.xhAlign}{The horizontal alignment type of the x-axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{plot.xvAlign}{The vertical alignment type of the x-axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{plot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is avaiable on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{plot.xLabelSize}{X-axis label size. Default is \code{10}.}

\item{plot.xTickLblSize}{X-axis tick label size. Default is \code{10}.}

\item{plot.yLabelSize}{Y-axis label size. Default is \code{10}.}

\item{plot.yTickLblSize}{Y-axis tick label size. Default is \code{10}.}

\item{plot.legendSize}{Legend size. Default is \code{9}.}

\item{plot.Width}{Scoreplot width. Default is \code{170}.}

\item{plot.Height}{Scoreplot height. Default is \code{150}.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}
}
\value{
Returns a pdf file for scoreplot.
}
\description{
scoreplot function for PLS-DA models.
}
\details{
When \code{length(comps) == 1}, the function generates a scatter plot plotting sample vs score for the comp of interest. When \code{length(comps) == 2}, the function generates a scatter plot plotting the two comps of interest against each other. When \code{length(comps) > 2}, the function generates a multi-panel correlation scoreplot matrix for the comps of interest - might be slow if the there are many comps.
}
\examples{
\dontrun{
rbioClass_plsda_scoreplot(new_model, comps = c(1, 2, 3), plot.ellipse = TRUE)
}
}
