% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_plsda.R
\name{rbioClass_plsda_tuplot}
\alias{rbioClass_plsda_tuplot}
\title{rbioClass_plsda_tuplot}
\usage{
rbioClass_plsda_tuplot(
  object,
  comps = 1,
  multi_plot.ncol = length(comps),
  multi_plot.nrow = 1,
  multi_plot.legend.pos = "bottom",
  plot.rightsideY = TRUE,
  plot.sampleLabel.type = "none",
  plot.sampleLabel.vector = NULL,
  plot.sampleLabelSize = 2,
  plot.sampleLabel.padding = 0.5,
  plot.SymbolSize = 5,
  plot.Title = NULL,
  plot.fontType = "sans",
  plot.xLabelSize = 10,
  plot.xTickLblSize = 10,
  plot.yLabelSize = 10,
  plot.yTickLblSize = 10,
  plot.legendSize = 9,
  plot.Width = 170,
  plot.Height = 150,
  verbose = TRUE
)
}
\arguments{
\item{object}{A \code{rbiomvr} object. Make sure the object is generated with a \code{validation} section.}

\item{comps}{Integer vector. Components to plot. The index of the components are integers. The vector length should be between 1 and the total number of components, inclusive. Can be Default is \code{c(1, 2)}.}

\item{multi_plot.ncol}{Set only when \code{length(comps) > 1}, number of columns on one figure page. Default is \code{length(comps)}.}

\item{multi_plot.nrow}{Set only when \code{length(comps) > 1}, number of rows on one figure page. Default is \code{1}.}

\item{multi_plot.legend.pos}{The legend position. Only effective when multi-plot is generated. Options are \code{"bottom"}, \code{"top"}, \code{"left"} and \code{"right"}. Default is \code{"bottom"}.}

\item{plot.rightsideY}{If to show the right side y-axis. Only applicable when the length of \code{comps} is less than 2, inclusive. Default is \code{FALSE}. Note: the right side Y is ignored when \code{length(comps) > 1}}

\item{plot.sampleLabel.type}{If to show the sample labels on the graph. Options are \code{"none"}, \code{"direct"} and \code{"indirect"}. Default is \code{"none"}.}

\item{plot.sampleLabel.vector}{Set only when \code{plot.sampleLabel.type} is not set to \code{"none"}, a character vector containing annotation (i.e. labels) for the samples. Default is \code{NULL}.}

\item{plot.sampleLabelSize}{Only set when \code{plot.sampleLabel.type} is not \code{"none"}. The size of the sample label. Default is \code{2}.}

\item{plot.sampleLabel.padding}{Set only when \code{plot.sampleLabel.type = "indirect"}, the padding between sample symbol and the label. Default is \code{0.5}.}

\item{plot.SymbolSize}{Symbol size. Default is \code{2}.}

\item{plot.Title}{tuplot title. Default is \code{NULL}.}

\item{plot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is available on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{plot.xLabelSize}{X-axis label size. Default is \code{10}.}

\item{plot.xTickLblSize}{X-axis tick label size. Default is \code{10}.}

\item{plot.yLabelSize}{Y-axis label size. Default is \code{10}.}

\item{plot.yTickLblSize}{Y-axis tick label size. Default is \code{10}.}

\item{plot.legendSize}{Legend size. Default is \code{9}.}

\item{plot.Width}{tuplot width. Default is \code{170}.}

\item{plot.Height}{tuplot height. Default is \code{150}.}

\item{verbose}{whether to display messages. Default is \code{TRUE}. This will not affect error or warning messages.}
}
\value{
Returns a pdf file for tuplot.
}
\description{
T-U plot function for PLS-DA models.
}
\details{
The T-U plot shows the correlation between the decomposed x and y matrices for PLS-DA analysis. Such plot is useful to inspecting X-Y decomposition and outlier detection. Since PLS-DA is a classification modelling method, the well-correlated T-U plot will likely show a logistic regression-like plot, as opposed to a linear regressoin plot. The \code{sampleLabel} series arguments make it possible to show exact sample, something useful for outlier detection. The function supports plotting multiple components at the same time, i.e. multiple plots on one page. The right side y-axis is not applicable when plotting multiple components.
}
\examples{
\dontrun{
rbioClass_plsda_tuplot(new_model, comps = c(1, 2, 3))
}
}
