% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_svm.R
\name{rbioClass_svm_cv}
\alias{rbioClass_svm_cv}
\title{rbioClass_svm_cv}
\usage{
rbioClass_svm_cv(
  x,
  y,
  center.scale = TRUE,
  kernel = c("radial", "linear", "polynomial", "sigmoid"),
  cross.k = 10,
  cross.best.model.method = c("none", "median"),
  tune.method = c("cross", "boot", "fix"),
  tune.cross.k = 10,
  tune.boot.n = 10,
  ...,
  parallelComputing = TRUE,
  n_cores = parallel::detectCores() - 1,
  clusterType = c("PSOCK", "FORK"),
  randomState = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{Input data matrix (e.g., independent variables, predictors, features, X, etc). Make sure it is either a matrix or a dataframe.}

\item{y}{Input response variable (e.g.,dependent variables, Y etc). Make sure it is \code{factor} class.}

\item{center.scale}{Logical, whether center and scale the data, i.e. subtracting mean (col.mean) and dividing by standard deviation (col.sd). Default is \code{TRUE}.}

\item{kernel}{SVM kernel. Options are \code{"linear", "ploynomial", "radial", "sigmoid"}. Default is \code{"radial"}, aka RBF.}

\item{cross.k}{Fold of nested cross validation, i.e. outer loop. Default is \code{10}.}

\item{cross.best.model.method}{The method to select the best cv models for feature selection. Options are \code{"median"} and \code{"none"}. Default is \code{"median"}.}

\item{tune.method}{Parameter tuning method, i.e. inner loop. Options are \code{"cross"} (i.e. cross validation), \code{"boot"} (i.e. bootstrap), and \code{"fix"}. Default is \code{"cross"}.}

\item{tune.cross.k}{Set only when \code{tune.method = "cross"}, fold number for cross validation. Default is \code{10}.}

\item{tune.boot.n}{Set only when \code{tune.method = "boot"}, bootstrap iterations. Default is \code{10}.}

\item{...}{Additional arguments for \code{rbioClass_svm}.}

\item{parallelComputing}{whether to use parallel computing or not. Default is \code{TRUE}.}

\item{n_cores}{Only set when \code{parallelComputing = TRUE}, the number of CPU cores to use. Default is \code{detectCores() - 1}, or the total number cores minus one.}

\item{clusterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}

\item{verbose}{whether to display messages. Default is \code{TRUE}. This will not affect error or warning messages.}
}
\value{
Returns a CV SVM model object: \code{rbiosvm_cv}.

Items for \code{rbiosvm_cv}:

     \code{univariate.fs}

     \code{cv.fold}: number of (outer) cross-validation fold

     \code{randomized.sample.index}: randomized sample order for (outer) cross-validation. \code{NULL} for classification modelling (cv.fold already randomized).

     \code{model.type}: the SVM model type, "classification" or "regression".

     \code{tot.accuracy.summary}: total (i.e. mean) nested cross-validation accuracy, if \code{model_type = "classification"}

     \code{tot.RMSE.summary}: total (i.e. mean) nested cross-validation RMSE, if \code{model_type = "regression"}

     \code{tot.rsq.summary}: total (i.e. mean) nested cross-validation R2, if \code{model_type = "regression"}

     \code{accuracy}: accuracy for each cross-validation iteration, if \code{model_type = "classification"}

     \code{RMSE}: RMSE for each cross-validation iteration, if \code{model_type = "regression"}

     \code{rsq}: R2 for each cross-validation iteration, if \code{model_type = "regression"}

     \code{cv.models}: all the CV SVM models

     \code{final.cv.models}: the CV SVM models selected by the \code{cross.best.model.method}.
                             when \code{cross.best.model.method = NULL}, \code{final.cv.models} is the same as \code{cv.models}.

     \code{nested.cv.models}: all the cv models with test data. NOTE: the test data is center-scaled with training data if \code{center.scale = TRUE}.

     \code{best.method}: method to determine the best model.

     \code{best.index}: index of the best models in the total model list

     \code{best.acc.summary}

     \code{best.rmse.summary}

     \code{best.rsq.summary}

     \code{best.accu}

     \code{best.rmse}

     \code{best.rsq}

     \code{tune.method}: (inner) loop method for SVM grid search

     \code{tune.cross.k}: fold number for (inner) loop if cross-validation is chosen

     \code{tune.boot.n}: iteration number for (inner) loop if bootstrap is chosen

     \code{run.time}: total run time
}
\description{
Cross-validation modelling and assessment for SVM modelling. It evaluates the overall performance of SVM modelling given the training data.
}
\details{
This function is very similar to \code{\link{rbioClass_svm_ncv_fs}}.
However, this function is intended for evaluating the final model, whereas the latter is for feature selection.

When \code{cross.best.model.method = "median"}, the function only use models with accuracy/RMSE equal or better than the median valaue
for feature count threshold. When there is no change in performance across cv models, the function behaves same as \code{cross.best.model.method = "none"}

The function also supports regression study, in which case, the performance metric is \code{RMSE}.

For classification modelling, the 10-fold CV segmentation is stratified, whereas the "normal" random resampling is used for regression modelling.
}
\examples{
\dontrun{
svm_cv <- rbioClass_svm_cv(x = mydata[, -c(1:2)],
                                     y = factor(mydata$Conditions, levels = unique(mydata$Conditions)),
                                     center.scale = TRUE,
                                     cross.k = 5,
                                     tune.method = "cross",
                                     tune.cross.k = 5,
                                     parallelComputing = TRUE, clusterType = "FORK")

}
}
