% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_svm_assessment.R
\name{rbioClass_svm_cv_roc_auc}
\alias{rbioClass_svm_cv_roc_auc}
\title{rbioClass_svm_cv_roc_auc}
\usage{
rbioClass_svm_cv_roc_auc(
  object,
  fileprefix = NULL,
  rocplot = TRUE,
  plot.smooth = FALSE,
  plot.lineSize = 1,
  plot.display.Title = TRUE,
  plot.titleSize = 10,
  plot.fontType = "sans",
  plot.xLabel = "1 - specificity",
  plot.xLabelSize = 10,
  plot.xTickLblSize = 10,
  plot.yLabel = "sensitivity",
  plot.yLabelSize = 10,
  plot.yTickLblSize = 10,
  plot.legendSize = 9,
  plot.rightsideY = TRUE,
  plot.Width = 170,
  plot.Height = 150,
  verbose = TRUE
)
}
\arguments{
\item{object}{A \code{rbiosvm_nestedcv} or \code{rbiosvm_cv} object.}

\item{fileprefix}{String. A file prefix to use for export file name, instead of the object name. Default is \code{NULL}.}

\item{rocplot}{If to generate a ROC plot. Default is \code{TRUE}.}

\item{plot.smooth}{If to smooth the curves. Uses binomial method to smooth the curves. Default is \code{FALSE}.}

\item{plot.lineSize}{Line size. Default is \code{1}.}

\item{plot.display.Title}{If to show the name of the y class. Default is \code{TRUE}.}

\item{plot.titleSize}{The font size of the plot title. Default is \code{10}.}

\item{plot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is avaiable on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{plot.xLabel}{X-axis label. Type with quotation marks. Could be NULL. Default is \code{"1 - specificity"}.}

\item{plot.xLabelSize}{X-axis label size. Default is \code{10}.}

\item{plot.xTickLblSize}{X-axis tick label size. Default is \code{10}.}

\item{plot.yLabel}{Y-axis label. Type with quotation marks. Could be NULL. Default is \code{"sensitivity"}.}

\item{plot.yLabelSize}{Y-axis label size. Default is \code{10}.}

\item{plot.yTickLblSize}{Y-axis tick label size. Default is \code{10}.}

\item{plot.legendSize}{Legend size. Default is \code{9}.}

\item{plot.rightsideY}{If to show the right side y-axis. Default is \code{FALSE}.}

\item{plot.Width}{Scoreplot width. Default is \code{170}.}

\item{plot.Height}{Scoreplot height. Default is \code{150}.}

\item{verbose}{whether to display messages. Default is \code{TRUE}. This will not affect error or warning messages.}

\item{plot.comps}{Number of comps to plot. Default is \code{1:object$ncomp}}

\item{plot.SymbolSize}{Symbol size. Default is \code{2}.}
}
\value{
Prints AUC values in the console. And a pdf file for ROC plot.
        The function also exports a ROC results list with an ROC object for each CV fold.
}
\description{
ROC-AUC analysis and ploting for SVM cross-validation (CV) models, for classification only.
}
\details{
Uses pROC module to calculate ROC. The function supports more than two groups or more than one threshold
         for classification and regression model.

         The test data included in the input \code{rbiosvm_nestedcv} is already normalized with training data information.
         So there is no need for additional data transformation.

         Also: a ROC curve with AUC == 1 is always 1-1 for 95% CI and can be misleading.
}
\examples{
\dontrun{
rbioClass_svm_cv_roc_auc(object = svm_nested_cv)
}
}
