% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_svm.R
\name{rbioClass_svm_cv_roc_auc_mean}
\alias{rbioClass_svm_cv_roc_auc_mean}
\title{rbioClass_svm_cv_roc_auc_mean}
\usage{
rbioClass_svm_cv_roc_auc_mean(
  object,
  fileprefix = NULL,
  rocplot = TRUE,
  plot.smooth = FALSE,
  plot.lineSize = 1,
  plot.display.Title = TRUE,
  plot.titleSize = 10,
  plot.fontType = "sans",
  plot.xLabel = "1 - specificity",
  plot.xLabelSize = 15,
  plot.xTickLblSize = 10,
  plot.yLabel = "sensitivity",
  plot.yLabelSize = 15,
  plot.yTickLblSize = 10,
  plot.legendSize = 9,
  plot.rightsideY = TRUE,
  plot.Width = 170,
  plot.Height = 150,
  verbose = TRUE
)
}
\arguments{
\item{object}{A \code{rbiosvm} object.}

\item{fileprefix}{String. A file prefix to use for export file name, instead of the objecte name. Default is \code{NULL}.}

\item{rocplot}{If to generate a ROC plot. Default is \code{TRUE}.}

\item{plot.smooth}{If to smooth the curves. Uses binormal method to smooth the curves. Default is \code{FALSE}.}

\item{plot.lineSize}{Line size. Default is \code{1}.}

\item{plot.display.Title}{If to show the name of the y class. Default is \code{TRUE}.}

\item{plot.titleSize}{The font size of the plot title. Default is \code{10}.}

\item{plot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is avaiable on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{plot.xLabel}{X-axis label. Type with quotation marks. Could be NULL. Default is \code{"1 - specificity"}.}

\item{plot.xLabelSize}{X-axis label size. Default is \code{10}.}

\item{plot.xTickLblSize}{X-axis tick label size. Default is \code{10}.}

\item{plot.yLabel}{Y-axis label. Type with quotation marks. Could be NULL. Default is \code{"sensitivity"}.}

\item{plot.yLabelSize}{Y-axis label size. Default is \code{10}.}

\item{plot.yTickLblSize}{Y-axis tick label size. Default is \code{10}.}

\item{plot.legendSize}{Legend size. Default is \code{9}.}

\item{plot.rightsideY}{If to show the right side y-axis. Default is \code{FALSE}.}

\item{plot.Width}{Scoreplot width. Default is \code{170}.}

\item{plot.Height}{Scoreplot height. Default is \code{150}.}

\item{verbose}{whether to display messages. Default is \code{TRUE}. This will not affect error or warning messages.}

\item{newdata}{A data matrix or vector for test data. Make sure it is a \code{matrix} or \code{vector} without labels, as well as the same feature numbers as the training set.}

\item{newdata.label}{The correspoding label vector to the data. Make sure it is a \code{factor} object. Defaults is \code{NULL}.}

\item{center.scale.newdata}{Logical, whether center and scale the newdata with training data mean and standard deviation. Default is \code{TRUE}.}

\item{plot.comps}{Number of comps to plot. Default is \code{1:object$ncomp}}

\item{plot.SymbolSize}{Symbol size. Default is \code{2}.}
}
\value{
1. PDF of the mean ROC-AUC plot
       2. A \code{v_rocauc_mean} to the environment
}
\description{
CV ROC-AUC mean analysis and ploting for SVM model, for classification model only.
}
\details{
The function construct a  list containing data.frames of AUC results over folds per group
       for every fold: list = roc dfm w cols: tpr, fpr, folds, auc dfm w cols: auc, folds

       The function also uses interpolation for CV ROC: set fpr seq(0, 1, length.out = 100), and interpolated tpr,
       to make sure all ROCs from different CV classifiers to have the same length
       interpolation for ROC thus is also used to plot ROCs for different types classifiers for comparison

       The ROC-AUC calcualte details follows the same as \code{\link{rbioClass_svm_roc_auc}}

       \code{v_rocauc_mean} items:
         1. \code{cv_auc_mean_res}: AUC calculation results
         2. \code{roc_plot}: mean ROC-AUC plot
}
