% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_svm.R
\name{rbioClass_svm_perm}
\alias{rbioClass_svm_perm}
\title{rbioClass_svm_perm()}
\usage{
rbioClass_svm_perm(
  object,
  perm.method = c("by_y", "by_feature_per_y"),
  nperm = 999,
  perm.plot = TRUE,
  ...,
  parallelComputing = TRUE,
  n_cores = parallel::detectCores() - 1,
  clusterType = c("PSOCK", "FORK"),
  verbose = TRUE
)
}
\arguments{
\item{object}{A \code{rbiosvm} object. Make sure the object is generated with a \code{tot.accuracy} section.}

\item{perm.method}{Permutation method. Options are \code{"by_y"} and \code{"by_feature_per_y"}. Default is \code{"by_y"}. See details below.}

\item{nperm}{Number of permutations to run. Default is \code{999}.}

\item{perm.plot}{Wether to produce a plot or not. Default is \code{TRUE}.}

\item{...}{Additional argument for \code{\link{rbioUtil_perm_plot}}.}

\item{parallelComputing}{Wether to use parallel computing or not. Default is \code{TRUE}.}

\item{n_cores}{Only set when \code{parallelComputing = TRUE}, the number of CPU cores to use. Default is \code{detectCores() - 1}, or the total number cores minus one.}

\item{clusterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}
}
\value{
The function returns \code{CSV} files for all intermediate permutation accuracy values as well as the p-value resutls.


The final results are also exported to the environment as a \code{rbiosvm_perm} object with the following items:

\code{perm.method} Permutation method.

\code{nperm} The number of permutation runs.

\code{performance.type} The stats metric used for the permutation test.

\code{original.performance} The original accuracy for the SVM model.

\code{perm.results} The intermediate permutation results, i.e. stats for each permutation test run in a data.frame. \code{nperm = 0} is the original stats.

\code{p.value} P value for the permutation test.

\code{model.type} SVM model type.

\code{run.time} run time

A scatter plot is also generaeted when \code{perm.plot = TRUE}.
}
\description{
Permutation test for SVM models.
}
\details{
The function uses RMSEP as the stats for comparing original model with permutatsions.

Data for permutation are object$centerX$centerX, meaning centered X are used in applicable.

Permutation methods are according to:

Ojala M, Garriga GC. 2010. Permutation test for studying classifier performance. J Mach Learn Res. 11: 1833 - 63.

For \code{perm.method = "by_y"}, labels (i.e. y) are permutatedted. A non-signifianct model (permutation p value > alpha, i.e. 0.05) in this case means the data is independent from the groups.

For \code{perm.method = "by_feature_per_by"}, X is first subset by label (i.e.y) before permutating data for each feature, i.e. by column. Since the permutation is done for the features WITHIN the group,
the test actually evaluates if the model will produce significantly different performannce from the permutation models with the original "betweeen-features" relation (if any) disturbed.
Therefore, A non-significant result (permutation p value > alpha, i.e. 0.05) means either the features are independent, or the model doesn't consider correlation between the features.
}
\examples{
\dontrun{
rbioClass_plsda_perm()
}
}
