% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_svm.R
\name{rbioClass_svm_predict}
\alias{rbioClass_svm_predict}
\title{rbioClass_svm_predict}
\usage{
rbioClass_svm_predict(
  object,
  newdata,
  center.scale.newdata = TRUE,
  export.name = NULL,
  sampleID.vector = NULL,
  newdata.y = NULL,
  prob.method = c("logistic", "Bayes", "softmax"),
  verbose = TRUE
)
}
\arguments{
\item{object}{A \code{rbiosvm} object.}

\item{newdata}{Input data to be classified. Make sure it is a \code{matrix} class and has the same variables as the model, i.e. same number of columns as the training data.}

\item{center.scale.newdata}{If to center the newdata. When \code{TRUE}, it will also apply the same scaling option as the \code{object}. Default is \code{TRUE}.}

\item{export.name}{String. Optional user defined export name prefix. Default is \code{NULL}.}

\item{sampleID.vector}{A character vector containing annotation (i.e. labels) for the samples. Default is \code{NULL}.}

\item{newdata.y}{Only required for regression study. The default is \code{NULL}.}

\item{prob.method}{Method to calculate classification probability. Options are \code{"logistic"}, \code{"softmax"} and \code{"Bayes"}. See details for more information. Default is \code{"logistic"}.}

\item{verbose}{whether to display messages. Default is \code{TRUE}. This will not affect error or warning messages.}
}
\value{
A \code{prediction} object. The items of the object are:

\code{model.type}

\code{classifier.class}

\code{predited.value}

\code{tot.predict.RMSE}

\code{prob.method}: Method to calculate posterior probability

\code{probability.summary}

\code{raw.newdata}

\code{newdata.id}

\code{center.scale}

\code{center.scaled.newdata}

\code{newdata.y}
}
\description{
Prediction function for SVM analysis. The function calculates the predicted value for unknown sample data using the input SVM model.
}
\details{
Although optional, the \code{newdata} matrix should be centered prior to testing, with the same scaling setting as the input \code{rbiosvm} object.

         The option \code{center.scale.newdata = FALSE} is for the already centered the data matrix. This center.scale process should use training data's
         column mean and column standard deviation.

         The default posterior probability calculation method \code{"logistic"} is the \code{e1071mc} package's implementation of logistic regression model.
         See \code{\link{rbioClass_plsda_predict}} for description for "Bayes" and "softmax" method.

         If \code{sampleID.vector = NULL} or missing, the function uses row numbers as label.

         When the model type is \code{"regression"}, the value of the irrelevant items is set to \code{NULL}. Likewise, when the model type is \code{"classification"},
         \code{newdata.y} and \code{tot.predict.RMSE} are set to \code{NULL}
}
\examples{
\dontrun{


}
}
