% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_svm.R
\name{rbioClass_svm_roc_auc}
\alias{rbioClass_svm_roc_auc}
\title{rbioClass_svm_roc_auc()}
\usage{
rbioClass_svm_roc_auc(object, newdata, newdata.label,
  center.scale.newdata = TRUE, rocplot = TRUE, plot.smooth = FALSE,
  plot.SymbolSize = 2, plot.display.Title = TRUE,
  plot.titleSize = 10, plot.fontType = "sans",
  plot.xLabel = "1 - specificity", plot.xLabelSize = 10,
  plot.xTickLblSize = 10, plot.yLabel = "sensitivity",
  plot.yLabelSize = 10, plot.yTickLblSize = 10, plot.legendSize = 9,
  plot.rightsideY = TRUE, plot.Width = 170, plot.Height = 150,
  verbose = TRUE)
}
\arguments{
\item{object}{A \code{rbiosvm} object.}

\item{newdata}{A data matrix or vector for test data. Make sure it is a \code{matrix} or \code{vector} without labels, as well as the same feature numbers as the training set.}

\item{newdata.label}{The correspoding label vector to the data. Make sure it is a \code{factor} object.}

\item{center.scale.newdata}{Logical, wether center and scale the newdata with training data mean and standard deviation. Default is \code{TRUE}.}

\item{rocplot}{If to generate a ROC plot. Default is \code{TRUE}.}

\item{plot.smooth}{If to smooth the curves. Uses binormal method to smooth the curves. Default is \code{FALSE}.}

\item{plot.SymbolSize}{Symbol size. Default is \code{2}.}

\item{plot.display.Title}{If to show the name of the y class. Default is \code{TRUE}.}

\item{plot.titleSize}{The font size of the plot title. Default is \code{10}.}

\item{plot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is avaiable on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{plot.xLabel}{X-axis label. Type with quotation marks. Could be NULL. Default is \code{"1 - specificity"}.}

\item{plot.xLabelSize}{X-axis label size. Default is \code{10}.}

\item{plot.xTickLblSize}{X-axis tick label size. Default is \code{10}.}

\item{plot.yLabel}{Y-axis label. Type with quotation marks. Could be NULL. Default is \code{"sensitivity"}.}

\item{plot.yLabelSize}{Y-axis label size. Default is \code{10}.}

\item{plot.yTickLblSize}{Y-axis tick label size. Default is \code{10}.}

\item{plot.legendSize}{Legend size. Default is \code{9}.}

\item{plot.rightsideY}{If to show the right side y-axis. Default is \code{FALSE}.}

\item{plot.Width}{Scoreplot width. Default is \code{170}.}

\item{plot.Height}{Scoreplot height. Default is \code{150}.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}

\item{plot.comps}{Number of comps to plot. Default is \code{1:object$ncomp}}
}
\value{
Prints AUC values in the console. And a pdf file for ROC plot. The function also exports a ROC results list to the environment.
}
\description{
ROC-AUC analysis and ploting for SVM model
}
\details{
Uses pROC module to calculate ROC.

Although optional, the \code{newdata} matrix should use training data's column mean and column standard deviation to enter.scale prior to ROC-AUC analysis.
The option \code{center.scaled.newdata = FALSE} is used when the whole (training and test sets) data were center.scaled before SVM training and testing.
}
\examples{
\dontrun{
rbioClass_plsda_roc_auc(object = model_binary, rocplot = TRUE, plot.comps = 1:2)
}
}
