% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioFS_app.R
\name{rbioFS_app}
\alias{rbioFS_app}
\title{rbioFS_app}
\usage{
rbioFS_app()
}
\description{
The web app version of \code{\link{rbioFS}}.
}
\examples{
\dontrun{
rbioFS_app() # launch the app version by running the function
}
}
