% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioFS_PCA.R
\name{rbioFS_PCA.file}
\alias{rbioFS_PCA.file}
\title{rbioFS_PCA.file}
\usage{
rbioFS_PCA.file(file, ...)
}
\arguments{
\item{file}{Input file name. Format is \code{csv}.}

\item{...}{Argument for \code{\link{rbioFS_PCA}}.}
}
\value{
Outputs a PCA object, a boxplot (proportion of variance) and a biplot from PCA analysis. The format is \code{pdf}.
}
\description{
A simple to use wrapper for PCA (Principal Component Analysis) and visualization. This is the version that loads \code{csv} file directly.
}
\details{
Make sure to arrange input data with first two columns for sample ID and conditions, and the rest for features (e.g., genes).
}
\examples{
\dontrun{
rbioFS_PCA.file(file = "data.csv", ellipse = TRUE, loadingPlot = TRUE, biplotWidth = 200, biplotHeight = 170)
}
}
