% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioFS_PCA.R
\name{rbioFS_PCA}
\alias{rbioFS_PCA}
\title{rbioFS_PCA}
\usage{
rbioFS_PCA(
  input = NULL,
  export.name = NULL,
  export.mode = c("full", "compatibility"),
  sampleIDVar = NULL,
  groupIDVar = NULL,
  scaleData = TRUE,
  centerData = TRUE,
  ...,
  boxplot = TRUE,
  boxplot.Title = NULL,
  boxplot.Width = 170,
  boxplot.Height = 150,
  biplot = TRUE,
  biplot.comps = c(1:2),
  biplot.Title = NULL,
  biplot.sampleLabel.type = c("none", "direct", "indirect"),
  biplot.sampleLabelSize = 2,
  biplot.sampleLabel.padding = 0.5,
  biplot.SymbolSize = 2,
  biplot.ellipse = FALSE,
  biplot.ellipse_conf = 0.95,
  biplot.xAngle = 0,
  biplot.xhAlign = 0.5,
  biplot.xvAlign = 0.5,
  biplot.loadingplot = FALSE,
  biplot.loadingplot.textsize = 3,
  biplot.mtx.densityplot = FALSE,
  biplot.mtx.stripLblSize = 10,
  biplot.Width = 170,
  biplot.Height = 150,
  rightsideY = FALSE,
  fontType = "sans",
  xTickLblSize = 10,
  yTickLblSize = 10,
  verbose = TRUE
)
}
\arguments{
\item{input}{Input data, data frame.}

\item{export.name}{String. Prefix for results exports. When \code{NULL}, the function uses the input object name. Default is \code{NULL}.}

\item{export.mode}{String. Export mode. Options are \code{full} and \code{compatibility}. Default is \code{full}.}

\item{sampleIDVar}{Sample variable name. It's a character string.}

\item{groupIDVar}{Group variable name. It's a character string.}

\item{scaleData}{If to scale the data when performing PCA. Default is \code{TRUE}.}

\item{centerData}{If to center the data when performing PCA. Default is \code{TRUE}.}

\item{...}{Additional arguments passed to \code{prcomp} function.}

\item{boxplot}{If to plot a boxplot for variance distribution. Default is \code{TRUE}.}

\item{boxplot.Title}{The boxplot title. Default is \code{NULL}.}

\item{boxplot.Width}{The boxplot width. Default is \code{170}.}

\item{boxplot.Height}{The boxplot height. Default is \code{150}.}

\item{biplot}{If to plot a scoreplot, with the option of superimpose a loading plot, i.e. biplot. Default is \code{TRUE}.}

\item{biplot.comps}{Integer or vector of integers. Index number(s) for principal component(s) to plot. Default is \code{c(1, 2)}.}

\item{biplot.Title}{The biplot title. Default is \code{NULL}.}

\item{biplot.sampleLabel.type}{If to show the sample labels on the graph. Options are \code{"none"}, \code{"direct"} and \code{"indirect"}. Default is \code{"none"}.}

\item{biplot.sampleLabelSize}{Only set when \code{biplot.sampleLabel.type} is not \code{"none"}, The size of the sample label. Default is \code{2}.}

\item{biplot.sampleLabel.padding}{Set only when \code{biplot.sampleLabel.type = "indirect"}, the padding between sample symbol and the label. Default is \code{0.5}.}

\item{biplot.SymbolSize}{The symbol size for the scatter plot portion of the biplot. Default is \code{2}.}

\item{biplot.ellipse}{If to draw ellipses. Default is \code{FALSE}.}

\item{biplot.ellipse_conf}{The confidence value for the ellipses. Default is \code{0.93}.}

\item{biplot.xAngle}{The rotation angle (degrees) of the biplot x axis marks. Default is \code{0} - horizontal.}

\item{biplot.xhAlign}{The horizontal alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{biplot.xvAlign}{The vertical alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{biplot.loadingplot}{If to superimpose loading plot. Default is \code{FALSE}.}

\item{biplot.loadingplot.textsize}{The font size of the loading plot labels. Default is \code{3}.}

\item{biplot.mtx.densityplot}{If to display a density plot on the diagonal for the correlation scoreplot matrix. Default is \code{FALSE}.}

\item{biplot.mtx.stripLblSize}{The label font size for the correlation scoreplot matrix strips. Default is \code{10}.}

\item{biplot.Width}{The biplot width. Default is \code{170}.}

\item{biplot.Height}{The biplot height. Default is \code{150}.}

\item{rightsideY}{If to show the right side y-axis for both boxplot and biplot. For biplot, only applicable when the length of \code{comps} is less than 2, inclusive. Default is \code{FALSE}.}

\item{fontType}{Font for the figure texts. Default is \code{"sans"}.}

\item{xTickLblSize}{X-axis tick label size. Default is \code{10}.}

\item{yTickLblSize}{Y-axis tick label size. Default is \code{10}.}

\item{verbose}{Whether to display messages. Default is \code{TRUE}. This will not affect error or warning messages.}
}
\value{
If \code{export.mode = "compatibility"}, then the function outputs a PCA object with suffix "_pca", a boxplot (proportion of variance) and a biplot from PCA analysis. The format is \code{pdf}.
       If \code{export.mode = "full"}, in addition to the PCA object exported by the "compatibility" mode, the function also exports a \code{"rbiofs_pca"} object, with suffix "pca_obj"

       Items of the \code{"rbio_pca"} object:
         input_dfm: input data.frame
         data_processing_settins: centre and scale settings
         scoreplot_dfm: score plot data.frame
         loadingvalueplot_dfm: loading plot data.frame, regardless of the \code{biplot.loadingplot} setting.
         boxplot_dfm: boxplot data.frame, regardless of the \code{boxplot} setting.
         pca: the PCA object
}
\description{
A simple to use wrapper for PCA (Principal Component Analysis) and visualization
}
\details{
Make sure to arrange input data with first two columns for sample ID and conditions, and the rest for features (e.g., genes).
}
\examples{
\dontrun{
rbioFS_PCA(input = pcaDfm, sampleIDVar = "conditions", groupIDVar = "samples",
biplot.ellipse = TRUE, biplot.loadingplot = TRUE, biplot.Width = 200, biplot.Height = 170)
}
}
