% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioFS_PCA_app.R
\name{rbioFS_PCA_app}
\alias{rbioFS_PCA_app}
\title{rbioFS_PCA_app}
\usage{
rbioFS_PCA_app()
}
\description{
The web app version of \code{\link{rbioFS_PCA}}.
}
\examples{
\dontrun{
rbioFS_PCA_app() # launch the app version by running the function
}
}
