% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioFS_plsda.R
\name{rbioFS_plsda}
\alias{rbioFS_plsda}
\title{rbioFS_plsda}
\usage{
rbioFS_plsda(x, y, ncomp = length(unique(y)) - 1, method = "simpls",
  scale = TRUE, validation = c("none", "CV", "LOO"), segments = 10,
  segments.type = "random", jackknife = TRUE, ...)
}
\arguments{
\item{x}{Input data matrix (e.g., independent variables, predictors, features, X, etc). Make sure it is either a matrix or a dataframe.}

\item{y}{Input response variable (e.g.,dependent variables, Y etc). Make sure it is \code{factor} class.}

\item{ncomp}{Number of components to be used for modelling. Default is \code{length(unique(y)) - 1}.}

\item{method}{PLS-DA modelling method. Four PLSR algorithms are available: the kernel algorithm ("kernelpls"), the wide kernel algorithm ("widekernelpls"), SIMPLS ("simpls") and the classical orthogonal scores algorithm ("oscorespls"). Default is the popular \code{"simpls}.}

\item{scale}{Logical, whether to scale the data or not. Default is \code{TRUE}.}

\item{validation}{Cross validation methods. Options are "none", "CV" (fold), "LOO" (leave-one-out). Default is \code{"CV"}.}

\item{segments}{Set only when \code{validation = "CV}, the number of segement to be set. Default is \code{10}.}

\item{segments.type}{Method to set up the segments. Options are \code{"random", "consecutive", "interleaved"}.Default is \code{"random"}.}

\item{jackknife}{If to use jack-knife procedure. Default is \code{TRUE}.}

\item{...}{Additional arguments for \code{mvr} function from \code{pls} pacakge.}
}
\value{
Returns a PLS-DA model object, with class "mvr".
}
\description{
PLS-DA modelling
}
\details{
For \code{ncomp} value, the default (full model) compares feature number with \code{class number - 1}, instead of observation number. For sequencing data, the input x needs to be either tranformed with the function like \code{clr_ilr_transfo()} from \code{RBioArray} package, or normalized using methods like "TMM" or "RLE" implemented in \code{edgeR} pacakge.
}
\examples{
\dontrun{
rbioFS_plsda(x, y, ncomp = 20)
}
}
