% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioFS_plsda.R
\name{rbioFS_plsda_classification}
\alias{rbioFS_plsda_classification}
\title{rbioFS_plsda_classification}
\usage{
rbioFS_plsda_classification(pred.obj, prob.method = "softmax",
  classplot = TRUE, multi_plot.ncol = nrow(pred.obj), multi_plot.nrow = 1,
  multi_plot.legend.pos = "bottom", multi_plot.stripLblSize = 10,
  plot.Title = NULL, plot.titleSize = 10, plot.probLabelSize = 2,
  plot.probLabel.padding = 0, plot.fontType = "sans", plot.legendSize = 9,
  plot.Width = 170, plot.Height = 150)
}
\arguments{
\item{prob.method}{The method used to calcuate classification probability. Options are "softmax" and "bayesian". Default is \code{"softmax"}.}

\item{classplot}{If to generate a classification pie plot. Default is \code{TRUE}.}

\item{multi_plot.ncol}{Number of columns on one figure page. Default is \code{nrow(pred.obj)}.}

\item{multi_plot.nrow}{Number of rows on one figure page. Default is \code{1}.}

\item{multi_plot.legend.pos}{The legend position. Only effective when multi-plot is generated. Options are \code{"bottom"}, \code{"top"}, \code{"left"} and \code{"right"}. Default is \code{"bottom"}.}

\item{plot.Title}{Plot title. Default is \code{FALSE}.}

\item{plot.titleSize}{The font size of the plot title. Default is \code{10}.}

\item{plot.probLabelSize}{The size of the sample label. Default is \code{2}.}

\item{plot.probLabel.padding}{Set only when \code{plot.sampleLabel.type = "indirect"}, the padding between sample symbol and the label. Default is \code{0.5}.}

\item{plot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is avaiable on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{plot.legendSize}{Legend size. Default is \code{9}.}

\item{plot.Width}{Scoreplot width. Default is \code{170}.}

\item{plot.Height}{Scoreplot height. Default is \code{150}.}

\item{pred.object}{A \code{prediction} object, which can be obtained from funciton \code{\link{rbioFS_plsda_predict}}.}

\item{plot.SymbolSize}{Symbol size. Default is \code{2}.}
}
\value{
A \code{classification} obejct with classification probability summary for each sample, as well as pdf figure file fif \code{classplot = TRUE}.
}
\description{
Classification function for PLS-DA analysis. The function calculates probability for each classification using the predicted value for the sample (i.e. test or unkown) data generated from \code{\link{rbioFS_plsda_predict}}.
}
\details{
The function operates in conjunction with the prediction function \code{\link{rbioFS_plsda_predict}}, to which the sample(s) of intestested is provided.

Based on the calculation, sample(s) will be classfied to the class with the highest classification probability. The \code{"softmax"} method calculates the classification using the equation below:
\code{probability = exp(predicted.value) / sum(exp(predicted.values))}

The Bayesian method is currently under development, and will be implemented in the next update.
}
\examples{
\dontrun{
ribioFS_plsda_classification(pred.obj = new_model_optm_plsda_predict, multi_plot.ncol = 4, multi_plot.nrow = 4, plot.probLabelSize = 2)
}
}
