% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioFS_plsda.R
\name{rbioFS_plsda_jackknife}
\alias{rbioFS_plsda_jackknife}
\title{rbioFS_plsda_jackknife}
\usage{
rbioFS_plsda_jackknife(object, ncomp = object$ncomp, use.mean = FALSE,
  sig.p = 0.05, plot = TRUE, plot.title = FALSE, plot.titleSize = 10,
  outlineCol = "black", errorbar = "SEM", errorbarWidth = 0.2,
  errorbarLblSize = 6, fontType = "sans", xLabel = NULL,
  xLabelSize = 10, xTickLblSize = 10, xTickItalic = FALSE,
  xTickBold = FALSE, xAngle = 0, xAlign = 0.5, rightsideY = TRUE,
  yLabel = NULL, yLabelSize = 10, yTickLblSize = 10,
  yTickItalic = FALSE, yTickBold = FALSE, legendSize = 9,
  legendTtl = FALSE, legendTtlSize = 9, plotWidth = 170,
  plotHeight = 150)
}
\arguments{
\item{object}{A \code{mvr} object. Make sure the object is generated with a \code{validation} section.}

\item{ncomp}{Defaults is all the components the \code{mvr} object has.}

\item{use.mean}{Defaults is \code{FALSE}.}

\item{sig.p}{Alpha value for the jack-knife coffecient p values. Defaults is \code{0.05}.}

\item{plot.title}{Whether to display plot title on top of the plot. Default is \code{FALSE}.}

\item{plot.titleSize}{The font size of the plot title. Default is \code{10}.}

\item{outlineCol}{The outline colour for the bar gars. Default is \code{"black"}.}

\item{errorbar}{Set the type of errorbar. Options are standard error of the mean (\code{"SEM"}, \code{"standard error"}, \code{"standard error of the mean"}), or standard deviation (\code{"SD"}, \code{"standard deviation"}), case insensitive. Default is \code{"SEM"}.}

\item{errorbarWidth}{Set the width for errorbar. Default is \code{0.2}.}

\item{errorbarLblSize}{Set the label size for the errorbar. Default is \code{6}.}

\item{fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is avaiable on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{xLabel}{x axis label. Type with quotation marks. Default is \code{NULL}.}

\item{xLabelSize}{x axis label size. Default is \code{10}.}

\item{xTickLblSize}{Font size of x axis ticks. Default is \code{10}.}

\item{xTickItalic}{Set x axis tick font to italic. Default is \code{FALSE}.}

\item{xTickBold}{Set x axis tick font to bold. Default is \code{FALSE}.}

\item{xAngle}{The rotation angle (degrees) of the x axis marks. Default is \code{0} - horizontal.}

\item{xAlign}{The alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{rightsideY}{If to display the right side y-axis. Default is \code{TRUE}.}

\item{yLabel}{y axis label. Type with quotation marks. Default is \code{NULL}.}

\item{yLabelSize}{y axis label size. Default is \code{10}.}

\item{yTickLblSize}{Font size of y axis ticks. Default is \code{10}.}

\item{yTickItalic}{Set y axis tick font to italic. Default is \code{FALSE}.}

\item{yTickBold}{Set y axis tick font to bold. Default is \code{FALSE}.}

\item{legendSize}{Legend size. Default is \code{9}.}

\item{legendTtl}{Hide/Display legend title. If \code{TRUE} or \code{T}, the name of the first column of the raw data file will display as the legend title. Default is \code{FALSE}.}

\item{legendTtlSize}{Set when \code{legendTtl = TRUE}, font size of the legend title. Default is \code{9}.}

\item{plotWidth}{The width of the plot (unit: mm). Default is 170. Default will fit most of the cases.}

\item{plotHeight}{The height of the plot (unit: mm). Default is 150. Default will fit most of the cases.}

\item{greyScale}{To set the graph in grey scale. Default is \code{TRUE}.}

\item{y_custom_tick_range}{To initiate setting the custom \code{y_upper_limit}, \code{y_lower_limit}, \code{y_major_tick_range}, \code{y_n_minor_ticks}. Default is \code{FALSE}.}

\item{y_upper_limit}{Can only be set when \code{y_custom_tick_range = TRUE}. Set custom upper limt for y axis. Value can be obtained from \code{\link{autorange_bar_y}}.}

\item{y_lower_limit}{Can only be set when \code{y_custom_tick_range = TRUE}. Set custom lower limt for y axis. Default is \code{0}. Value can be obtained from \code{\link{autorange_bar_y}}.}

\item{y_major_tick_range}{Can only be set when \code{y_custom_tick_range = TRUE}. Set custom major tick range for y axis.  Value can be obtained from \code{\link{autorange_bar_y}}.}

\item{y_n_minor_ticks}{Can only be set when \code{y_custom_tick_range = TRUE}. Set custom numbers of minor ticks. Default is \code{4}. Value can be obtained from \code{\link{autorange_bar_y}}.}
}
\value{
Outputs two list objects to the environment, one for raw Jack-Knife results matrices, one for plot dataframe. Also the function also generates the pdf figure files to the working directory.
}
\description{
Jack-Knife procedure for the \code{PLS} models, e.g. \code{PLS-DA} or \coee{PLS-R}.
}
\details{
\code{use.mean = FALSE} is more main stream. Make sure to use cross validated and optimized component number for \code{ncomp}.
}
\examples{
\dontrun{
rbioFS(file = "test.csv", impute = TRUE, imputeIter = 50, quantileNorm = TRUE)
}
}
