% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioFS_plsda.R
\name{rbioFS_plsda_ncomp_select}
\alias{rbioFS_plsda_ncomp_select}
\title{rbioFS_plsda_ncomp_select()}
\usage{
rbioFS_plsda_ncomp_select(object, ..., ncomp.selection.method = "1sd",
  randomization.nperm = 999, randomization.alpha = 0.05, rmsepplot = TRUE,
  rightsideY = TRUE, rmsepplot.optm.ncomp.line = TRUE,
  multi_rmsepplot.ncol = length(dimnames(object$coefficients)[[2]]),
  multi_rmsepplot.nrow = 1, rmsepplot.display.Title = TRUE,
  rmsepplot.SymbolSize = 2, rmsepplot.fontType = "sans",
  rmsepplot.xTickLblSize = 10, rmsepplot.yTickLblSize = 10,
  rmsepplot.Width = 170, rmsepplot.Height = 150)
}
\arguments{
\item{object}{A \code{rbiomvr} or \code{mvr} object. Make sure the object is generated with a \code{validation} section.}

\item{...}{Additional argument for \code{RMSEP} function from \code{pls} package.}

\item{ncomp.selection.method}{Optimal numbers of components selection method. Options are \code{"min"}, \code{"1sd"}, and \code{"randomization"}. Default is \code{"1sd"}.}

\item{randomization.nperm}{Set only when \code{ncomp.selection.method = "randomization"}, number of permutations. Default is \code{999}.}

\item{randomization.alpha}{Set only when \code{ncomp.selection.method = "randomization"}, alpha for the p values used during "randomization" selection. Default is \code{0.05}.}

\item{rmsepplot}{If to generate a RMSEP plot. Default is \code{TRUE}.}

\item{rightsideY}{If to show the right side y-axis. Only applicble when the length of \code{comps} is less than 2, inclusive. Default is \code{FALSE}. Note: the right side Y is ignored when \code{length(comps) > 1}}

\item{rmsepplot.optm.ncomp.line}{If to display the vertical line indicting the optimal number of components. Default is \code{TRUE}.}

\item{multi_rmsepplot.ncol}{Number of columns on one figure page. Default is the number of responding classes, i.e. y.}

\item{multi_rmsepplot.nrow}{Number of rows on one figure page. Default is \code{1}.}

\item{rmsepplot.display.Title}{If to show the name of the y class. Default is \code{TRUE}.}

\item{rmsepplot.SymbolSize}{Symbol size. Default is \code{2}.}

\item{rmsepplot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is avaiable on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{rmsepplot.xTickLblSize}{X-axis tick label size. Default is \code{10}.}

\item{rmsepplot.yTickLblSize}{Y-axis tick label size. Default is \code{10}.}

\item{rmsepplot.Width}{Scoreplot width. Default is \code{170}.}

\item{rmsepplot.Height}{Scoreplot height. Default is \code{150}.}
}
\value{
Prints the selected number of components for each y class. Returns RMSEP values for each y class to the environment, as well as a pdf file for the RMSEP plot if \code{rmsepplot = TRUE}.
}
\description{
Optimal number of components selection for PLS-DA model, with RMSEP plot funcitonality. Selection methods are modified based on \code{selectNcomp()} from \code{pls} pacakge.
}
\details{
Three methods are used for components number selection: \code{"min"} simply chooses the number of components to reach te minimum RMSEP; \code{"1sd"} chooses the number of components when its RMSEP first reaches minimum as well as within one standard deviation; For "randomization", see the help file for \code{selectNcomp()} function from  \code{pls} pacakge.
}
\examples{
\dontrun{
rbioFS_plsda_ncomp_select(new_model,  multi_rmsepplot.ncol = 2, multi_rmsepplot.nrow = 2, min.rmsep.line = T,
         ncomp.selection.method = "randomization", randomization.nperm = 999, randomization.alpha = 0.05)
}
}
