% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioFS_plsda.R
\name{rbioFS_plsda_q2r2}
\alias{rbioFS_plsda_q2r2}
\title{rbioFS_plsda_q2r2()}
\usage{
rbioFS_plsda_q2r2(object, intercept = TRUE, q2r2plot = TRUE,
  plot.display.Title = TRUE,
  multi_plot.ncol = length(dimnames(object$coefficients)[[2]]),
  multi_plot.nrow = 1, multi_plot.legend.pos = "bottom",
  plot.rightsideY = TRUE, plot.fontType = "sans", plot.SymbolSize = 2,
  plot.xLabel = "Components", plot.xLabelSize = 10,
  plot.xTickLblSize = 10, plot.yLabel = "R2 & Q2", plot.yLabelSize = 10,
  plot.yTickLblSize = 10, plot.legendSize = 9, plot.Width = 170,
  plot.Height = 150, verbose = TRUE)
}
\arguments{
\item{object}{A \code{rbiomvr} or \code{mvr} object. Make sure the object is generated with a \code{validation} section.}

\item{intercept}{Wether to include intercept term, i.e. comps = 0. Default is \code{TRUE}.}

\item{q2r2plot}{If to generate a q2r2 plot. Default is \code{TRUE}.}

\item{plot.display.Title}{If to show the name of the y class. Default is \code{TRUE}.}

\item{multi_plot.ncol}{Number of columns on one figure page. Default is the number of responding classes, i.e. unique y classes.}

\item{multi_plot.nrow}{Number of rows on one figure page. Default is \code{1}.}

\item{multi_plot.legend.pos}{The legend position. Only effective when multi-plot is generated. Options are \code{"bottom"}, \code{"top"}, \code{"left"} and \code{"right"}. Default is \code{"bottom"}.}

\item{plot.rightsideY}{If to show the right side y-axis. Default is \code{FALSE}. Note: doesn't seem to be necessasry as PLS-DA always has at least two y classes.}

\item{plot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is avaiable on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{plot.SymbolSize}{Symbol size. Default is \code{2}.}

\item{plot.xLabel}{X-axis label. Type with quotation marks. Could be NULL. Default is \code{"Components"}.}

\item{plot.xLabelSize}{X-axis label size. Default is \code{10}.}

\item{plot.xTickLblSize}{X-axis tick label size. Default is \code{10}.}

\item{plot.yLabel}{Y-axis label. Type with quotation marks. Could be NULL. Default is \code{"R2 & Q2"}.}

\item{plot.yLabelSize}{Y-axis label size. Default is \code{10}.}

\item{plot.yTickLblSize}{Y-axis tick label size. Default is \code{10}.}

\item{plot.legendSize}{Legend size. Default is \code{9}.}

\item{plot.Width}{Scoreplot width. Default is \code{170}.}

\item{plot.Height}{Scoreplot height. Default is \code{150}.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}.}
}
\value{
Prints the selected number of components for each y class. Returns RMSEP values for each y class to the environment, as well as a pdf file for the RMSEP plot if \code{rmsepplot = TRUE}.
}
\description{
q2-r2 (i.e. Q^2 and R^2 scores) caluclation and plot for plsda models
}
\details{
A vertical line indicating the number of component with minimum q2-r2 distance.
}
\examples{
\dontrun{
rbioFS_plsda_q2r2(object = new_model, multi_plot.ncol = 2, multi_plot.nrow = 2, intercept = TRUE)
}
}
