% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioFS_plsda.R
\name{rbioFS_plsda_scoreplot}
\alias{rbioFS_plsda_scoreplot}
\title{rbioFS_plsda_scoreplot}
\usage{
rbioFS_plsda_scoreplot(object, comps = c(1, 2), rightsideY = FALSE,
  scoreplot.Title = NULL, scoreplot.SymbolSize = 2,
  scoreplot.ellipse = FALSE, scoreplot.ellipse_conf = 0.95,
  cor.scoreplot.densityplot = FALSE, cor.scoreplot.stripLblSize = 10,
  scoreplot.xAngle = 0, scoreplot.xhAlign = 0.5, scoreplot.xvAlign = 0.5,
  scoreplot.fontType = "sans", scoreplot.xTickLblSize = 10,
  scoreplot.yTickLblSize = 10, scoreplot.Width = 170,
  scoreplot.Height = 150)
}
\arguments{
\item{object}{A \code{mvr} object. Make sure the object is generated with a \code{validation} section.}

\item{comps}{Integer vector. Components to plot. The index of the components are intergers. The vector length should be between 1 and the total number of components, inclusive. Can be Default is \code{c(1, 2)}.}

\item{rightsideY}{If to show the right side y-axis. Only applicble when the length of \code{comps} is less than 2, inclusive. Default is \code{FALSE}.}

\item{scoreplot.Title}{Scoreplot title. Default is \code{NULL}.}

\item{scoreplot.SymbolSize}{Symbol size. Default is \code{2}.}

\item{scoreplot.ellipse}{If to draw ellipses. Default is \code{FALSE}.}

\item{scoreplot.ellipse_conf}{The confidence value for the ellipses. Default is \code{0.95}.}

\item{cor.scoreplot.densityplot}{If to display a density plot on the diagonal for the correlation scoreplot matrix. Default is \code{FALSE}.}

\item{cor.scoreplot.stripLblSize}{The label font size for the correlation scoreplot matrix strips. Default is \code{10}.}

\item{scoreplot.xAngle}{The rotation angle (degrees) of the x axis marks. Default is \code{0} - horizontal.}

\item{scoreplot.xhAlign}{The horizontal alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{scoreplot.xvAlign}{The vertical alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{scoreplot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is avaiable on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{scoreplot.xTickLblSize}{X-axis tick label size. Default is \code{10}.}

\item{scoreplot.yTickLblSize}{Y-axis tick label size. Default is \code{10}.}

\item{scoreplot.Width}{Scoreplot width. Default is \code{170}.}

\item{scoreplot.Height}{Scoreplot height. Default is \code{150}.}
}
\value{
Returns a pdf file for scoreplot.
}
\description{
scoreplot function for PLS-DA models.
}
\details{
When \code{length(comps) == 1}, the function generates a scatter plot plotting sample vs score for the comp of interest. When \code{length(comps) == 2}, the function generates a scatter plot plotting the two comps of interest against each other. When \code{length(comps) > 2}, the function generates a multi-panel correlation scoreplot matrix for the comps of interest - might be slow if the there are many comps.
}
\examples{
\dontrun{
rbioFS_plsda_scoreplot(new_model, comps = c(1, 2, 3), scoreplot.ellipse = TRUE)
}
}
