% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioFS_plsda.R
\name{rbioFS_plsda_VIP}
\alias{rbioFS_plsda_VIP}
\title{rbioFS_plsda_VIP}
\usage{
rbioFS_plsda_VIP(object, vip.alpha = 1, plot = TRUE, plot.title = TRUE,
  plot.titleSize = 10, plot.sig.line = TRUE, plot.outlineCol = "black",
  plot.errorbar = "SEM", plot.errorbarWidth = 0.2,
  plot.errorbarLblSize = 6, plot.fontType = "sans",
  plot.xLabel = "Features", plot.xLabelSize = 10, plot.xTickLblSize = 10,
  plot.xTickItalic = FALSE, plot.xTickBold = FALSE, plot.xAngle = 90,
  plot.xhAlign = 0.95, plot.xvAlign = 0.5, plot.rightsideY = TRUE,
  plot.yLabel = "VIP", plot.yLabelSize = 10, plot.yTickLblSize = 10,
  plot.yTickItalic = FALSE, plot.yTickBold = FALSE, plot.legendSize = 9,
  plot.legendTtl = FALSE, plot.legendTtlSize = 9, plot.Width = 170,
  plot.Height = 150, verbose = TRUE)
}
\arguments{
\item{object}{A \code{mvr} or \code{rbiomvr} object. Make sure the object is generated with a \code{validation} section.}

\item{vip.alpha}{Alpha value (threshold) for VIP values. Any VIP above this is considered important. Defaults is \code{1}.}

\item{plot.title}{Whether to display plot title on top of the plot. Default is \code{FALSE}.}

\item{plot.titleSize}{The font size of the plot title. Default is \code{10}.}

\item{plot.sig.line}{Wether to display a horizontal line indicating the VIP threshold. Default is \code{TRUE}.}

\item{plot.outlineCol}{The outline colour for the bar gars. Default is \code{"black"}.}

\item{plot.errorbar}{Set the type of errorbar. Only applicable if the object model is built using more than one component. Options are standard error of the mean (\code{"SEM"}, \code{"standard error"}, \code{"standard error of the mean"}), or standard deviation (\code{"SD"}, \code{"standard deviation"}), case insensitive. Default is \code{"SEM"}.}

\item{plot.errorbarWidth}{Set the width for errorbar. Default is \code{0.2}.}

\item{plot.errorbarLblSize}{Set the label size for the errorbar. Default is \code{6}.}

\item{plot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is avaiable on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{plot.xLabel}{X-axis label. Type with quotation marks. Could be NULL. Default is \code{"Features"}.}

\item{plot.xLabelSize}{X-axis label size. Default is \code{10}.}

\item{plot.xTickLblSize}{Font size of x axis ticks. Default is \code{10}.}

\item{plot.xTickItalic}{Set X-axis tick font to italic. Default is \code{FALSE}.}

\item{plot.xTickBold}{Set X-axis tick font to bold. Default is \code{FALSE}.}

\item{plot.xAngle}{The rotation angle (degrees) of the x axis marks. Default is \code{0} - horizontal.}

\item{plot.xhAlign}{The horizontal alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{plot.xvAlign}{The vertical alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{plot.rightsideY}{If to display the right side y-axis. Default is \code{TRUE}.}

\item{plot.yLabel}{Y-axis label. Type with quotation marks. Could be NULL. Default is \code{"VIP"}.}

\item{plot.yLabelSize}{Y-axis label size. Default is \code{10}.}

\item{plot.yTickLblSize}{Font size of y axis ticks. Default is \code{10}.}

\item{plot.yTickItalic}{Set Y-axis tick font to italic. Default is \code{FALSE}.}

\item{plot.yTickBold}{Set Y-axis tick font to bold. Default is \code{FALSE}.}

\item{plot.Width}{The width of the plot (unit: mm). Default is 170. Default will fit most of the cases.}

\item{plot.Height}{The height of the plot (unit: mm). Default is 150. Default will fit most of the cases.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}.}
}
\value{
Outputs a list objects to the environment with the VIP raw values, VIP summary and the ncomp value. Also the function also generates the pdf figure files to the working directory.
}
\description{
VIP, or variable importance in projection, calcualtion and plotting for plsda models. This is another FS method, and can be used independently.
}
\details{
Only works when the plsda model is fitted with the orthorgonal score algorithm, or NIPALS. Such model can be built using \code{\link{rbioFS_plsda}} with \code{method = "oscorespls"}.
         For each feature, the boxplot is the mean of the VIP values from all the components, hence with errorbars. However, if the model is fitted using only one component, the function will automatically adjust.
         The VIP threshold of 1 is the most commonly accepted value. However it is also acceptable to set according to the data and objectives of the study.
}
\examples{
\dontrun{
rbioFS_plsda_VIP(object = new_model_optm,
                       vip.alpha = 0.05,
                       plot = TRUE, plot.title = TRUE, plot.titleSize = 10,
                       plot.sig.line = TRUE,
                       plot.outlineCol = "black", plot.errorbar = "SEM", plot.errorbarWidth = 0.2,
                       plot.errorbarLblSize = 6, plot.fontType = "sans", plot.xLabel = "Features",
                       plot.xLabelSize = 10, plot.xTickLblSize = 10, plot.xTickItalic = FALSE,
                       plot.xTickBold = FALSE, plot.xAngle = 90, plot.xhAlign = 1, plot.xvAligh = 0.2, plot.rightsideY = TRUE,
                       plot.yLabel = "Coefficients", plot.yLabelSize = 10, plot.yTickLblSize = 10,
                       plot.yTickItalic = FALSE, plot.yTickBold = FALSE, plot.legendSize = 9,
                       plot.legendTtl = FALSE, plot.legendTtlSize = 9, plot.Width = 170,
                       plot.Height = 150)
}
}
