% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_plsda.R
\name{rbioFS_plsda_vip_plot}
\alias{rbioFS_plsda_vip_plot}
\title{rbioFS_plsda_vip_plot}
\usage{
rbioFS_plsda_vip_plot(
  vip_obj,
  plot.preview = TRUE,
  plot.title = TRUE,
  plot.titleSize = 10,
  plot.sig.line = TRUE,
  plot.outlineCol = "black",
  plot.errorbar = "SEM",
  plot.errorbarWidth = 0.2,
  plot.errorbarLblSize = 6,
  plot.fontType = "sans",
  plot.xLabel = "Features",
  plot.xLabelSize = 10,
  plot.xTickLblSize = 10,
  plot.xTickItalic = FALSE,
  plot.xTickBold = FALSE,
  plot.xAngle = 90,
  plot.xhAlign = 0.95,
  plot.xvAlign = 0.5,
  plot.rightsideY = TRUE,
  plot.yLabel = "VIP",
  plot.yLabelSize = 10,
  plot.yTickLblSize = 10,
  plot.yTickItalic = FALSE,
  plot.yTickBold = FALSE,
  plot.legendSize = 9,
  plot.legendTtl = FALSE,
  plot.legendTtlSize = 9,
  plot.Width = 170,
  plot.Height = 150,
  verbose = TRUE
)
}
\arguments{
\item{vip_obj}{A \code{rbiomvr_vip}, generated by \code{\link{rbioFS_plsda_VIP}}.}

\item{plot.preview}{If to preview plots. Default is \code{TRUE}.}

\item{plot.title}{Whether to display plot title on top of the plot. Default is \code{FALSE}.}

\item{plot.titleSize}{The font size of the plot title. Default is \code{10}.}

\item{plot.sig.line}{whether to display a horizontal line indicating the VIP threshold. Default is \code{TRUE}.}

\item{plot.outlineCol}{The outline colour for the bar graph. Default is \code{"black"}.}

\item{plot.errorbar}{Set the type of errorbar. Only applicable if the object model is built with bootstrapping. Options are standard error of the mean (\code{"SEM"}, \code{"standard error"}, \code{"standard error of the mean"}), or standard deviation (\code{"SD"}, \code{"standard deviation"}), case insensitive. Default is \code{"SEM"}.}

\item{plot.errorbarWidth}{Set the width for errorbar. Default is \code{0.2}.}

\item{plot.errorbarLblSize}{Set the label size for the errorbar. Default is \code{6}.}

\item{plot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is available on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{plot.xLabel}{X-axis label. Type with quotation marks. Could be NULL. Default is \code{"Features"}.}

\item{plot.xLabelSize}{X-axis label size. Default is \code{10}.}

\item{plot.xTickLblSize}{Font size of x axis ticks. Default is \code{10}.}

\item{plot.xTickItalic}{Set X-axis tick font to italic. Default is \code{FALSE}.}

\item{plot.xTickBold}{Set X-axis tick font to bold. Default is \code{FALSE}.}

\item{plot.xAngle}{The rotation angle (degrees) of the x axis marks. Default is \code{0} - horizontal.}

\item{plot.xhAlign}{The horizontal alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{plot.xvAlign}{The vertical alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{plot.rightsideY}{If to display the right side y-axis. Default is \code{TRUE}.}

\item{plot.yLabel}{Y-axis label. Type with quotation marks. Could be NULL. Default is \code{"VIP"}.}

\item{plot.yLabelSize}{Y-axis label size. Default is \code{10}.}

\item{plot.yTickLblSize}{Font size of y axis ticks. Default is \code{10}.}

\item{plot.yTickItalic}{Set Y-axis tick font to italic. Default is \code{FALSE}.}

\item{plot.yTickBold}{Set Y-axis tick font to bold. Default is \code{FALSE}.}

\item{plot.Width}{The width of the plot (unit: mm). Default is 170. Default will fit most of the cases.}

\item{plot.Height}{The height of the plot (unit: mm). Default is 150. Default will fit most of the cases.}

\item{verbose}{whether to display messages. Default is \code{TRUE}. This will not affect error or warning messages.}
}
\value{
Outputs pdf figure files to the working directory.
}
\description{
Plotting function for \code{\link{rbioFS_plsda_VIP}}.
}
\details{
Only works with \code{rbiomvr_vip} objects. Set \code{plot.preview = FALSE} to speed up the process as preview rendering may be slow.
         The function also applies to the regression model (model.type = "regression").
}
\examples{
\dontrun{
rbioFS_plsda_vip_plot(vip_obj = plsda_m_vip,
                      plot.title = TRUE, plot.titleSize = 10,
                      plot.sig.line = TRUE,
                      plot.outlineCol = "black", plot.errorbar = "SEM", plot.errorbarWidth = 0.2,
                      plot.errorbarLblSize = 6, plot.fontType = "sans", plot.xLabel = "Features",
                      plot.xLabelSize = 10, plot.xTickLblSize = 10, plot.xTickItalic = FALSE,
                      plot.xTickBold = FALSE, plot.xAngle = 90, plot.xhAlign = 1, plot.xvAligh = 0.2, plot.rightsideY = TRUE,
                      plot.yLabel = "Coefficients", plot.yLabelSize = 10, plot.yTickLblSize = 10,
                      plot.yTickItalic = FALSE, plot.yTickBold = FALSE, plot.legendSize = 9,
                      plot.legendTtl = FALSE, plot.legendTtlSize = 9, plot.Width = 170,
                      plot.Height = 150)
}
}
