% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioRF_initialFS.R
\name{rbioFS_rf_initialFS}
\alias{rbioFS_rf_initialFS}
\title{rbioFS_rf_initialFS}
\usage{
rbioFS_rf_initialFS(objTitle = "data", x, y, nTimes = 50,
  nTree = 1001, mTry = if (!is.factor(y)) max(floor(ncol(x)/3), 1) else
  floor(sqrt(ncol(x))), parallelComputing = TRUE,
  n_cores = parallel::detectCores() - 1, clusterType = "PSOCK",
  plot = TRUE, n = "all", ...)
}
\arguments{
\item{objTitle}{The title for the output data frame. Default is \code{"x_vs_tgt"}}

\item{x}{Input dataframe or matrix. Make sure to arrange the data with features as column names. Note this excludes the label column.}

\item{y}{The target (outcome) variable for random forest feature selection.}

\item{nTimes}{Number of random forest vi computation runs. Default is \code{50} times.}

\item{nTree}{Number of trees generated for each random forest run. Default is \code{1001} trees.}

\item{mTry}{Number of random feature pick when building the tree. Default is \code{max(floor(ncol(dfm) / 3), 2)} for classification, and \code{floor(sqrt(ncol(x)))}.}

\item{parallelComputing}{Wether to use parallel computing or not. Default is \code{TRUE}.}

\item{n_cores}{Only set when \code{parallelComputing = TRUE}, the number of CPU cores to use. Default is \code{detectCores() - 1}, or the total number cores minus one.}

\item{clusterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}

\item{plot}{If to plot a bargraph to visualize vi and the ranking. Default is \code{TRUE}}

\item{n}{Number of features to show. Takes integer numbers. Default is \code{"all"} (make sure to include quotation marks).}

\item{...}{Additional arguments passed to the plot function \code{\link{rbioFS_rf_initialFS_plot}}.}
}
\value{
Outputs a \code{rf_ifs} object with vi values for each feature and OOB error rate. When \code{TRUE}, bargraph for the vi is also generated and exported as a \code{.pdf} file.
}
\description{
Recursive random froest variable importance (vi) and OOB error rate computation.
}
\details{
Make sure to arrange data (dfm) with feature (e.g., gene) as variables (i.e., columns), and rownames as sample names.
}
\examples{
\dontrun{
rbioFS_rf_initialFS(training_HCvTC, tgtVar_HCvTC, n = 40, errorbar = "SEM", plotWidth = 400, plotHeight = 200)
}
}
