% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioFS_rf_sfs.R
\name{rbioFS_rf_SFS}
\alias{rbioFS_rf_SFS}
\title{rbioFS_rf_SFS}
\usage{
rbioFS_rf_SFS(objTitle = "x_vs_tgt", x, y, nTimes = 50, nTree = 1001,
  parallelComputing = TRUE, n_cores = parallel::detectCores() - 1,
  clusterType = "PSOCK", plot = TRUE, n = "all", ...)
}
\arguments{
\item{objTitle}{The title for the output data frame. Default is \code{"x_vs_tgt"}}

\item{x}{Input dataframe or matrix. Make sure to arrange the data with features as column names. Note this excludes the label column.}

\item{y}{The target (outcome) variable for random forest feature selection.}

\item{nTimes}{Number of random forest vi computation runs. Default is \code{50} times.}

\item{nTree}{Number of trees generated for each random forest run. Default is \code{1001} trees.}

\item{parallelComputing}{Wether to use parallel computing or not. Default is \code{TRUE}.}

\item{n_cores}{Only set when \code{parallelComputing = TRUE}, the number of CPU cores to use. Default is \code{detectCores() - 1}, or the total number cores minus one.}

\item{clusterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}

\item{plot}{If to plot a bargraph to visualize vi and the ranking. Default is \code{TRUE}}

\item{n}{Number of features to show. Takes integer numbers. Default is \code{"all"} (make sure to include quotation marks).}

\item{...}{Additional arguments passed to the plot function \code{\link{rbioFS_rf_SFS_plot}}.}
}
\value{
Outputs a \code{rf_sfs} object with  OOB error rate summary, and a joint-point curve in \code{csv} format.
}
\description{
Recursive nested random froest variable importance (vi) and OOB error rate computation in a sequential forward selection (SFS) manner.
}
\details{
Make sure to arrange data (dfm) with feature (e.g., gene) as variables (i.e., columns), and rownames as sample names.
}
\examples{
\dontrun{
rbioFS_rf_SFS(training_HCvTC, tgtVar_HCvTC, parallelComputing = TRUE)
}
}
