% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioFS_rf_sfs.R
\name{rbioFS_rf_SFS_plot}
\alias{rbioFS_rf_SFS_plot}
\title{rbioFS_rf_SFS_plot}
\usage{
rbioFS_rf_SFS_plot(object, n = "all",
  plot.file.title = deparse(substitute(object)), plot.title = NULL,
  plot.titleSize = 10, plot.symbolSize = 2, plot.errorbar = "sem",
  plot.errorbarWidth = 0.2, plot.fontType = "sans",
  plot.xLabel = "Features", plot.xLabelSize = 10, plot.xAngle = 0,
  plot.xhAlign = 0.5, plot.xvAlign = 0.5, plot.xTickLblSize = 10,
  plot.xTickItalic = FALSE, plot.xTickBold = FALSE,
  plot.rightsideY = TRUE, plot.yLabel = "OOB error rate",
  plot.yLabelSize = 10, plot.yTickLblSize = 10,
  plot.yTickItalic = FALSE, plot.yTickBold = FALSE, plot.Width = 170,
  plot.Height = 150, verbose = TRUE)
}
\arguments{
\item{object}{A \code{rf_sfs} object. Make sure the object is generated with a \code{validation} section.}

\item{n}{Number of features to show. Takes integer numbers. Default is \code{"all"} (make sure to include quotation marks).}

\item{plot.file.title}{Plot file title prefix. Defatuls is \code{deparse(substitute(object))}.}

\item{plot.title}{Whether to display plot title on top of the plot. Default is \code{FALSE}.}

\item{plot.titleSize}{The font size of the plot title. Default is \code{10}.}

\item{plot.errorbar}{Set the type of errorbar. Options are standard error of the mean (\code{"SEM"}, \code{"standard error"}, \code{"standard error of the mean"}), or standard deviation (\code{"SD"}, \code{"standard deviation"}), case insensitive. Default is \code{"SEM"}.}

\item{plot.errorbarWidth}{Set the width for errorbar. Default is \code{0.2}.}

\item{plot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is avaiable on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{plot.xLabel}{x axis label. Type with quotation marks. Could be NULL. Default is \code{"Features"}.}

\item{plot.xLabelSize}{x axis label size. Default is \code{10}.}

\item{plot.xAngle}{The rotation angle (degrees) of the x axis marks. Default is \code{0} - horizontal.}

\item{plot.xhAlign}{The horizontal alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{plot.xvAlign}{The vertical alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{plot.xTickLblSize}{Font size of x axis ticks. Default is \code{10}.}

\item{plot.xTickItalic}{Set x axis tick font to italic. Default is \code{FALSE}.}

\item{plot.xTickBold}{Set x axis tick font to bold. Default is \code{FALSE}.}

\item{plot.rightsideY}{If to display the right side y-axis. Default is \code{TRUE}.}

\item{plot.yLabel}{y axis label. Type with quotation marks. Could be NULL. Default is \code{"OOB error rate"}.}

\item{plot.yLabelSize}{y axis label size. Default is \code{10}.}

\item{plot.yTickLblSize}{Font size of y axis ticks. Default is \code{10}.}

\item{plot.yTickItalic}{Set y axis tick font to italic. Default is \code{FALSE}.}

\item{plot.yTickBold}{Set y axis tick font to bold. Default is \code{FALSE}.}

\item{plot.Width}{The width of the plot (unit: mm). Default is 170. Default will fit most of the cases.}

\item{plot.Height}{The height of the plot (unit: mm). Default is 150. Default will fit most of the cases.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}

\item{plot.SymbolSize}{Symbol size. Default is \code{2}.}
}
\value{
The function also generates the pdf error curve figure files to the working directory.
}
\description{
Variable error curve plot function for \code{rf_sfs} obejcts generated by \code{rbioFS_rf_SFS()} or \code{rbioFS()}.
}
\examples{
\dontrun{
rbioFS_rf_SFS_plot(object = data_SFS, n = "all",plot.title = TRUE, plot.titleSize = 10,
                       plot.outlineCol = "black", plot.errorbar = "SEM", plot.errorbarWidth = 0.2,
                       plot.errorbarLblSize = 6, plot.fontType = "sans", plot.xLabel = "Features",
                       plot.xLabelSize = 10, plot.xTickLblSize = 10, plot.xTickItalic = FALSE,
                       plot.xTickBold = FALSE, plot.xAngle = 90, plot.xhAlign = 1, plot.xvAligh = 0.2, plot.rightsideY = TRUE,
                       plot.yLabel = "Mean Decrease in Accuracy", plot.yLabelSize = 10, plot.yTickLblSize = 10,
                       plot.yTickItalic = FALSE, plot.yTickBold = FALSE, plot.legendSize = 9,
                       plot.legendTtl = FALSE, plot.legendTtlSize = 9, plot.Width = 170,
                       plot.Height = 150)
}
}
