% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioFS_svm.R
\name{rbioFS_svm_roc_auc}
\alias{rbioFS_svm_roc_auc}
\title{rbioFS_svm_roc_auc()}
\usage{
rbioFS_svm_roc_auc(object, newdata, newdata.label, center.newdata = TRUE,
  scale.newdata = TRUE, rocplot = TRUE, plot.smooth = FALSE,
  plot.rightsideY = TRUE, plot.SymbolSize = 2, plot.display.Title = TRUE,
  plot.titleSize = 10, plot.fontType = "sans",
  plot.xLabel = "1 - specificity", plot.xLabelSize = 10,
  plot.xTickLblSize = 10, plot.yLabel = "sensitivity",
  plot.yLabelSize = 10, plot.yTickLblSize = 10, plot.legendSize = 9,
  plot.Width = 170, plot.Height = 150, verbose = TRUE)
}
\arguments{
\item{object}{A \code{rbiosvm} object.}

\item{newdata}{A data matrix or vector for test data. Make sure it is a \code{matrix} or \code{vector} without labels, as well as the same feature numbers as the training set.}

\item{newdata.label}{The correspoding label vector to the data. Make sure it is a \code{factor} object.}

\item{center.newdata}{Logical, wether center the newdata, i.e. subtracting mean. Default is \code{TRUE}.}

\item{scale.newdata}{Logical, whether to scale the newdata, i.e. dividing by standard deviation. Default is \code{TRUE}.}

\item{rocplot}{If to generate a ROC plot. Default is \code{TRUE}.}

\item{plot.smooth}{If to smooth the curves. Uses binormal method to smooth the curves. Default is \code{FALSE}.}

\item{plot.rightsideY}{If to show the right side y-axis. Default is \code{FALSE}. Note: doesn't seem to be necessasry as PLS-DA always has at least two y classes.}

\item{plot.SymbolSize}{Symbol size. Default is \code{2}.}

\item{plot.display.Title}{If to show the name of the y class. Default is \code{TRUE}.}

\item{plot.titleSize}{The font size of the plot title. Default is \code{10}.}

\item{plot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is avaiable on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{plot.xLabel}{X-axis label. Type with quotation marks. Could be NULL. Default is \code{"1 - specificity"}.}

\item{plot.xLabelSize}{X-axis label size. Default is \code{10}.}

\item{plot.xTickLblSize}{X-axis tick label size. Default is \code{10}.}

\item{plot.yLabel}{Y-axis label. Type with quotation marks. Could be NULL. Default is \code{"sensitivity"}.}

\item{plot.yLabelSize}{Y-axis label size. Default is \code{10}.}

\item{plot.yTickLblSize}{Y-axis tick label size. Default is \code{10}.}

\item{plot.legendSize}{Legend size. Default is \code{9}.}

\item{plot.Width}{Scoreplot width. Default is \code{170}.}

\item{plot.Height}{Scoreplot height. Default is \code{150}.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will be affect error or warning messeages.}

\item{plot.comps}{Number of comps to plot. Default is \code{1:object$ncomp}}
}
\value{
Prints AUC values in the console. And a pdf file for ROC plot. The function also exports a ROC results list to the environment.
}
\description{
ROC-AUC analysis and ploting for SVM model
}
\details{
Uses pROC module to calculate ROC.

Although optional, the \code{newdata} matrix should have the same center/scale settings as the SVM modelling prior to the ROC-AUC analysis.
The option \code{center.newdata = FALSE} and \code{scale.newdata = FALSE} can also be used for the already processed the newdata matrix.
}
\examples{
\dontrun{
rbioFS_plsda_roc_auc(object = model_binary, rocplot = TRUE, plot.comps = 1:2)
}
}
