% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileprocessing.R
\name{rbioNorm}
\alias{rbioNorm}
\title{rbioNorm}
\usage{
rbioNorm(RawData, NormMtd = "quantile", correctBG = TRUE,
  BgMtd = "normexp", BgOffst = 0)
}
\arguments{
\item{RawData}{Inpule dataframe, which can be the the results from the function \code{\link{rbioIMP}} with transposition.}

\item{NormMtd}{Normalization method. Options are \code{"none"}, \code{"scale"}, \code{"quantile"} and \code{"cyclicloess"}. Default is \code{"quantile"}. See \code{limma} package for details.}

\item{correctBG}{If to correct background. Default is \code{TRUE}.}

\item{BgMtd}{Background correction method if \code{correctBG} is TRUE. Options are \code{"auto"}, \code{"none"}, \code{"subtract"}, \code{"half"}, \code{"minimum"}, \code{"movingmin"}, \code{"edwards"} and \code{"normexp"}. Default is \code{"normexp"}. See \code{limma} package for details.}

\item{BgOffst}{Background offset if \code{correctBG} is TRUE. Default is \code{0}. Briefly, this is the constant value to add to the data. See \code{limma} package for details.}
}
\value{
Outputs a \code{dataframe} with processed data.
}
\description{
Data normalization. This is a shell function using limma package.
}
\examples{
\dontrun{
normdata <- rbioNorm(rawT, correctBG = FLASE)
}
}
