% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rboReg_plsr.R
\name{rbioReg_plsr_perm}
\alias{rbioReg_plsr_perm}
\title{rbioReg_plsr_perm()}
\usage{
rbioReg_plsr_perm(
  object,
  ncomp = object$ncomp,
  adjCV = FALSE,
  perm.plot = TRUE,
  ...,
  nperm = 999,
  parallelComputing = TRUE,
  n_cores = parallel::detectCores() - 1,
  clusterType = c("PSOCK", "FORK"),
  verbose = TRUE
)
}
\arguments{
\item{object}{A \code{rbiomvr} object. Make sure the object is generated with a \code{validation} section.}

\item{ncomp}{Model complexity, i.e. number of components to model. Default is \code{object$ncomp}, i.e. maximum complexity.}

\item{adjCV}{If to use adjusted CV, i.e. CV adjusted for unbalanced data. Default is \code{FALSE}.}

\item{perm.plot}{Whether to produce a plot or not. Default is \code{TRUE}.}

\item{...}{Additional argument for \code{\link{rbioUtil_perm_plot}}.}

\item{nperm}{Number of permutations to run. Default is \code{999}.}

\item{parallelComputing}{Whether to use parallel computing or not. Default is \code{TRUE}.}

\item{n_cores}{Only set when \code{parallelComputing = TRUE}, the number of CPU cores to use. Default is \code{detectCores() - 1}, or the total number cores minus one.}

\item{clusterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}

\item{verbose}{Whether to display messages. Default is \code{TRUE}. This will not affect error or warning messages.}
}
\value{
The function returns \code{CSV} files for all intermediate permutation RMSEP values as well as the p-value results.

The results are also exported to the environment as a \code{rbiomvr_perm} object with the following items:

\code{nperm} The number of permutation runs.

\code{perm.stats} The stats metric used for the permutation test.

\code{adjCV} If the adjusted cross-validation stats is used.

\code{p.value.summary} P values for the permutation test.

\code{poerm.results} The intermediate permutation results, i.e. stats for each permutation test run in a data.frame. \code{nperm = 0} is the original stats.

A scatter plot is also generated when \code{perm.plot = TRUE}.
}
\description{
Permutation test for PLSR models.
}
\details{
The function uses RMSEP as the stats for comparing original model with permutations.

         Usually, we use the optimized PLS-DA model for \code{object}, which can be obtained from functions \code{\link{rbioClass_plsda}} and \code{\link{rbioClass_plsda_ncomp_select}}.

         Data for permutation are object$centerX$centerX, meaning centered X are used if applicable.

         Permutation methods are according to:

            Ojala M, Garriga GC. 2010. Permutation test for studying classifier performance.
            J Mach Learn Res. 11: 1833 - 63.

         For this function, labels (i.e. y) are permutated. A non-significant model (permutation p value > alpha, i.e. 0.05) in this case means the data is independent from the groups.
}
\examples{
\dontrun{
rbioReg_plsr_perm(object = new_model_optm, nperm = 999, adjCV = TRUE, parallelComputing = TRUE)
}
}
