% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_svm_assessment.R
\name{rbioReg_svm_r2}
\alias{rbioReg_svm_r2}
\title{rbioReg_svm_r2}
\usage{
rbioReg_svm_r2(object, newdata = NULL, newdata.y = NULL)
}
\arguments{
\item{object}{A \code{rbiosvm} object.}

\item{newdata}{A data matrix or vector for test data. Make sure it is a \code{matrix} or \code{vector} without labels, as well as the same feature numbers as the training set.}

\item{newdata.y}{For regression model only, the vector for the new data's continuous outcome variable. Default is \code{NULL}}
}
\value{
RMSE value with either new data or training data.
}
\description{
Support Vector Regression (SVR) R2 calculation
}
\details{
The R2 is calculated as following: 1 - rss/tss.
rss: residual sum of squares: sum((yhat-y)^2)
tss: total sum of squares: sum((y-mean(y))^2)
}
\examples{
\dontrun{
 test_r2 <- rbioReg_svm_r2(object = svm_m, newdata = svm_test[,-1], newdata.y = svm_test$y)
}
}
