% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_svm_assessment.R
\name{rbioReg_svm_rmse}
\alias{rbioReg_svm_rmse}
\title{rbioReg_svm_rmse}
\usage{
rbioReg_svm_rmse(object, newdata = NULL, newdata.y = NULL)
}
\arguments{
\item{object}{A \code{rbiosvm} object.}

\item{newdata}{A data matrix or vector for test data. Make sure it is a \code{matrix} or \code{vector} without labels, as well as the same feature numbers as the training set.}

\item{newdata.y}{For regression model only, the vector for the new data's continuous outcome variable. Default is \code{NULL}}
}
\value{
RMSE value with either new data or training data.
}
\description{
Support Vector Regression (SVR) RMSE calculation
}
\details{
With newdata, the newdata is center_scaled using training data parameter (SD etc).
Without newdata, the RMSE is based on the CV RMSE.
}
\examples{
\dontrun{
 test_rmse <- rbioReg_svm_rmse(object = svm_m, newdata = svm_test[,-1], newdata.y = svm_test$y)
}
}
