% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioRF_initialFS.R
\name{rbioRF_initialFS}
\alias{rbioRF_initialFS}
\title{rbioRF_initialFS}
\usage{
rbioRF_initialFS(objTitle = "x_vs_tgt", x, targetVar, nTimes = 50,
  nTree = 1001, mTry = max(floor(ncol(x)/3), 2), multicore = TRUE,
  plot = TRUE, n = "all", Title = NULL,
  xLabel = "Mean Decrease in Accuracy", yLabel = NULL, errorbar = "SEM",
  errorbarWidth = 0.2, xTickLblSize = 10, yTickLblSize = 10,
  plotWidth = 170, plotHeight = 150)
}
\arguments{
\item{objTitle}{The title for the output data frame. Default is \code{"x_vs_tgt"}}

\item{x}{Input dataframe or matrix. Make sure to arrange the data with features as column names.}

\item{targetVar}{The target variable for random forest feature selection. This is a factor object.}

\item{nTimes}{Number of random forest vi computation runs. Default is \code{50} times.}

\item{nTree}{Number of trees generated for each random forest run. Default is \code{1001} trees.}

\item{mTry}{Number of random feature pick when building the tree. Default is \code{max(floor(ncol(dfm) / 3), 2)}.}

\item{multicore}{If to use parallel computing. Default is \code{TRUE}.}

\item{plot}{If to plot a bargraph to visualize vi and the ranking. Default is \code{TRUE}}

\item{n}{Number of features to show. Takes integer numbers. Default is \code{"all"} (make sure to include quotation marks).}

\item{Title}{Figure title. Make sure to use quotation marks. Use \code{NULL} to hide. Default is \code{NULL}.}

\item{xLabel}{X-axis label. Make sure to use quotation marks. Use \code{NULL} to hide. Default is \code{NULL}.}

\item{yLabel}{Y-axis label. Make sure to use quotatio marks. Use \code{NULL} to hide. Default is \code{"Mean Decrease in Accurac"}}

\item{errorbar}{The type of errorbar in the graph. Options are \code{"SEM"} (standard error of the mean) or \code{"SD"} (standard deviation). Default is \code{"SEM"}.}

\item{errorbarWidth}{The width of the errorbar. Default is \code{0.2}.}

\item{xTickLblSize}{Font size for the x-axis text. Default is \code{10}.}

\item{yTickLblSize}{Font size for the y-axis text. Default is \code{10}.}

\item{plotWidth}{The width of the figure for the final output figure file. Default is \code{170}.}

\item{plotHeight}{The height of the figure for the final output figure file. Default is \code{150}.}
}
\value{
Outputs a \code{list} object with vi values for each feature and OOB error rate. When \code{TRUE}, bargraph for the vi is also generated and exported as a \code{.pdf} file.
}
\description{
Recursive random froest variable importance (vi) and OOB error rate computation.
}
\details{
Make sure to arrange data (dfm) with feature (e.g., gene) as variables (i.e., columns), and rownames as sample names.
}
\examples{
\dontrun{
rbioRF_initialFS(training_HCvTC, tgtVar_HCvTC, n = 40, errorbar = "SEM", plotWidth = 400, plotHeight = 200)
}
}
