% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioUtil.R
\name{rbioUtil_classif_accuracy}
\alias{rbioUtil_classif_accuracy}
\title{Title rbioUtil_classif_accuracy}
\usage{
rbioUtil_classif_accuracy(
  object,
  newdata,
  newdata.label,
  center.scale.newdata = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{object}{Classification model object. The object should be either \code{rbiosvm}, or \code{rbiomvr} classes.}

\item{newdata}{A data matrix or vector for test data. Make sure it is a \code{matrix} or \code{vector} without labels, as well as the same feature numbers as the training set.}

\item{newdata.label}{Permutation test results object. The object should be either \code{rbiosvm_perm}, or \code{rbiomvr_perm} classes.}

\item{center.scale.newdata}{Logical, whether center and scale the newdata with training data mean and standard deviation. Default is \code{TRUE}.}
}
\value{
Numerical model accuracy value.
}
\description{
Plotting function for permutation test results.
}
\details{
accuracy = true predictions/total predictions
}
\examples{

\dontrun{
rbioUtil_classif_accuracy(perm_res = svm_model_perm)
}

}
