% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioUtil.R
\name{rbioUtil_classplot}
\alias{rbioUtil_classplot}
\title{rbioUtil_classplot}
\usage{
rbioUtil_classplot(
  pred.obj,
  multi_plot.ncol = nrow(pred.obj),
  multi_plot.nrow = 1,
  multi_plot.legend.pos = "bottom",
  multi_plot.stripLblSize = 10,
  plot.Title = NULL,
  plot.titleSize = 10,
  plot.probLabelSize = 5,
  plot.probLabel.padding = 0,
  plot.probLabel.outside = FALSE,
  plot.probLabel.outside.nudge = 1.5,
  plot.fontType = "sans",
  plot.legendSize = 9,
  plot.Width = 170,
  plot.Height = 150,
  verbose = TRUE
)
}
\arguments{
\item{multi_plot.ncol}{Number of columns on one figure page. Default is \code{nrow(pred.obj)}.}

\item{multi_plot.nrow}{Number of rows on one figure page. Default is \code{1}.}

\item{multi_plot.legend.pos}{The legend position. Only effective when multi-plot is generated. Options are \code{"bottom"}, \code{"top"}, \code{"left"} and \code{"right"}. Default is \code{"bottom"}.}

\item{plot.Title}{Plot title. Default is \code{FALSE}.}

\item{plot.titleSize}{The font size of the plot title. Default is \code{10}.}

\item{plot.probLabelSize}{The size of the sample label. Default is \code{2}.}

\item{plot.probLabel.padding}{Set only when \code{plot.sampleLabel.type = "indirect"}, the padding between sample symbol and the label. Default is \code{0.5}.}

\item{plot.probLabel.outside}{If to put the probability label outside of the pies. Default is \code{"TRUE"}.}

\item{plot.probLabel.outside.nudge}{Set only when \code{plot.probLabel.outside = TRUE}, adjustment to nudge the starting position of each label. Default is \code{"1.5"}.}

\item{plot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is available on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{plot.legendSize}{Legend size. Default is \code{9}.}

\item{plot.Width}{Scoreplot width. Default is \code{170}.}

\item{plot.Height}{Scoreplot height. Default is \code{150}.}

\item{verbose}{whether to display messages. Default is \code{TRUE}. This will not affect error or warning messages.}

\item{pred.object}{A \code{prediction} object, which can be obtained from function \code{\link{rbioClass_plsda_predict}}.}

\item{plot.SymbolSize}{Symbol size. Default is \code{2}.}
}
\value{
A \code{classification} object with classification probability summary for each sample, as well as pdf figure file when \code{classplot = TRUE}.
}
\description{
Classification plot function for classification prediction. The function uses \code{prediction} object generated from \code{\link{rbioClass_plsda_predict}} or \code{\link{rbioClass_svm_predict}} to generate classification probablity pie charts.
}
\details{
The function operates in conjunction with the prediction function \code{\link{rbioClass_plsda_predict}}, to which the sample(s) of interested is provided.
}
\examples{
\dontrun{
rbioUtil_classplot(pred.obj = new_model_optm_plsda_predict, multi_plot.ncol = 4, multi_plot.nrow = 4, plot.probLabelSize = 2)
}
}
