% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioUtil.R
\name{rbioUtil_fscount_plot}
\alias{rbioUtil_fscount_plot}
\alias{rbioUtil_fscount_plot.rbiosvm_nestedcv}
\alias{rbioUtil_fscount_plot.default}
\title{Title rbioUtil_fscount_plot}
\usage{
rbioUtil_fscount_plot(object, ...)

\method{rbioUtil_fscount_plot}{rbiosvm_nestedcv}(object, export.name, ...)

\method{rbioUtil_fscount_plot}{default}(
  dfm,
  feature_var,
  count_var,
  export.name = "plot",
  threshold = 2,
  plot.preview = TRUE,
  plot.title = TRUE,
  plot.titleSize = 10,
  plot.sig.line = TRUE,
  plot.outlineCol = "gray4",
  plot.fillCol = "lightcyan",
  plot.errorbarLblSize = 6,
  plot.fontType = "sans",
  plot.xLabel = "Features",
  plot.xLabelSize = 10,
  plot.xTickLblSize = 10,
  plot.xTickItalic = FALSE,
  plot.xTickBold = FALSE,
  plot.xAngle = 90,
  plot.xhAlign = 1,
  plot.xvAlign = 0.5,
  plot.rightsideY = TRUE,
  plot.yLabel = "FS counts",
  plot.yLabelSize = 10,
  plot.yTickLblSize = 10,
  plot.yTickItalic = FALSE,
  plot.yTickBold = FALSE,
  plot.legendSize = 9,
  plot.legendTtl = FALSE,
  plot.legendTtlSize = 9,
  plot.Width = 170,
  plot.Height = 150
)
}
\arguments{
\item{object}{A \code{rbiosvm_nestedcv} object.}

\item{...}{Additional arguments for the default method.}

\item{export.name}{Name used to output the plot.}

\item{dfm}{Input data.frame.}

\item{feature_var}{Variable name for feature names.}

\item{count_var}{Variable name for FS count.}

\item{threshold}{An integer indicating the FS count threshold for consensus feature selection.}

\item{plot.preview}{If to preview plots. Default is \code{TRUE}.}

\item{plot.title}{Whether to display plot title on top of the plot. Default is \code{FALSE}.}

\item{plot.titleSize}{The font size of the plot title. Default is \code{10}.}

\item{plot.outlineCol}{The outline colour for the bar graph. Default is \code{"gray4"}.}

\item{plot.fillCol}{The fill colour for the bar. Default is \code{"lightcyan"}}

\item{plot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is available on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{plot.xLabel}{X-axis label. Type with quotation marks. Could be NULL. Default is \code{"Features"}.}

\item{plot.xLabelSize}{X-axis label size. Default is \code{10}.}

\item{plot.xTickLblSize}{Font size of x axis ticks. Default is \code{10}.}

\item{plot.xTickItalic}{Set X-axis tick font to italic. Default is \code{FALSE}.}

\item{plot.xTickBold}{Set X-axis tick font to bold. Default is \code{FALSE}.}

\item{plot.xAngle}{The rotation angle (degrees) of the x axis marks. Default is \code{0} - horizontal.}

\item{plot.xhAlign}{The horizontal alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{plot.xvAlign}{The vertical alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{plot.rightsideY}{If to display the right side y-axis. Default is \code{TRUE}.}

\item{plot.yLabel}{Y-axis label. Type with quotation marks. Could be NULL. Default is \code{"VIP"}.}

\item{plot.yLabelSize}{Y-axis label size. Default is \code{10}.}

\item{plot.yTickLblSize}{Font size of y axis ticks. Default is \code{10}.}

\item{plot.yTickItalic}{Set Y-axis tick font to italic. Default is \code{FALSE}.}

\item{plot.yTickBold}{Set Y-axis tick font to bold. Default is \code{FALSE}.}

\item{plot.Width}{The width of the plot (unit: mm). Default is 170. Default will fit most of the cases.}

\item{plot.Height}{The height of the plot (unit: mm). Default is 150. Default will fit most of the cases.}
}
\value{
ggplot 2 bar graph exported as PDF file
}
\description{
Bar graph for the FS count numbers
}
\details{
The input data.frame should at least have variables for feature name and FS count.
}
