% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioUtil.R
\name{rbioUtil_onehotEncoding}
\alias{rbioUtil_onehotEncoding}
\title{rbioUtil_onehotEncoding}
\usage{
rbioUtil_onehotEncoding(data, cols, export_mode = c("dfm", "dt"))
}
\arguments{
\item{data}{data.frame. Input data.frame.}

\item{cols}{str or vector of strings. Column names to encode.}

\item{export_mode}{str. Export as data.table or data.frame. Default is \code{"dfm"}.}
}
\value{
data.frame with columns one hot encoded.
}
\description{
Convert data.frame features into factors.
}
\details{
The output data.frame has all the data of the inoput data, as opposed to just encoded columns.
}
\examples{
\dontrun{
  rbioUtil_onehotEncoding(dfm = data, cols = c("gender"), export_mode = "dfm")
}
}
