% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioUtil.R
\name{rbioUtil_perm_plot.default}
\alias{rbioUtil_perm_plot.default}
\title{Title rbioUtil_perm_plot.default}
\usage{
\method{rbioUtil_perm_plot}{default}(baseplt, plot.display.Title = TRUE,
  plot.titleSize = 10, plot.fontType = "sans",
  plot.xLabel = "Permutations", plot.xLabelSize = 10,
  plot.xTickLblSize = 10, plot.yLabel = "Accuracy",
  plot.yLabelSize = 10, plot.yTickLblSize = 10, plot.legendSize = 9,
  plot.rightsideY = TRUE)
}
\arguments{
\item{plot.display.Title}{If to show the name of the y class. Default is \code{TRUE}.}

\item{plot.titleSize}{The font size of the plot title. Default is \code{10}.}

\item{plot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is avaiable on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{plot.xLabel}{X-axis label. Type with quotation marks. Could be NULL. Default is \code{"Permutation"}.}

\item{plot.xLabelSize}{X-axis label size. Default is \code{10}.}

\item{plot.xTickLblSize}{X-axis tick label size. Default is \code{10}.}

\item{plot.yLabel}{Y-axis label. Type with quotation marks. Could be NULL.}

\item{plot.yLabelSize}{Y-axis label size. Default is \code{10}.}

\item{plot.yTickLblSize}{Y-axis tick label size. Default is \code{10}.}

\item{plot.legendSize}{Legend size. Default is \code{9}.}

\item{plot.rightsideY}{If to show the right side y-axis. Default is \code{FALSE}.}

\item{baseplot}{A base plot generated by the \code{\link{rbioUtil_perm_plot}} method.}
}
\value{
A baseplot object for permutation results.
}
\description{
Core plotting function for SVM permutation test results.
}
