% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioUtil.R
\name{rbioUtil_perm_plot.rbiomvr_perm}
\alias{rbioUtil_perm_plot.rbiomvr_perm}
\title{Title rbioUtil_perm_plot.rbiomvr_perm}
\usage{
\method{rbioUtil_perm_plot}{rbiomvr_perm}(perm_res, plot.SymbolSize = 2,
  plot.Width = 170, plot.Height = 150, ..., verbose = TRUE)
}
\arguments{
\item{perm_res}{Permutation test results object. The object should be \code{rbiomvr_perm} class.}

\item{plot.SymbolSize}{Symbol size. Default is \code{2}.}

\item{plot.Width}{Scoreplot width. Default is \code{170}.}

\item{plot.Height}{Scoreplot height. Default is \code{150}.}

\item{...}{Additional argument for the plot settings.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}
}
\value{
A pdf file containing a scatter plot for permutation results.
}
\description{
Plotting function for PLS-DA permutation test results.
}
