% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioUtil.R
\name{rbioUtil_perm_plot}
\alias{rbioUtil_perm_plot}
\title{Title rbioUtil_perm_plot}
\usage{
rbioUtil_perm_plot(perm_res, ...)
}
\arguments{
\item{perm_res}{Permutation test results object. The object should be either \code{rbiosvm_perm}, or \code{rbiomvr_perm} classes.}

\item{...}{Additional argument for the plot settings, see details.}
}
\value{
A pdf file containing a scatter plot for permutation results.
}
\description{
Generic plotting function for permutation test results.
}
\examples{

\dontrun{
rbioUtil_perm_plot(perm_res = svm_model_perm)
}

}
