% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioUtil.R
\name{shap_g_theme_helper}
\alias{shap_g_theme_helper}
\title{shap_g_theme_helper}
\usage{
shap_g_theme_helper(
  g,
  plot.SymbolSize = 2,
  plot.lineSize = 1,
  plot.display.Title = TRUE,
  plot.title = "SHAP",
  plot.titleSize = 10,
  plot.fontType = "sans",
  plot.xLabel = "SHAP value",
  plot.xLabelSize = 10,
  plot.xTickLblSize = 10,
  plot.yLabel = "Features",
  plot.yLabelSize = 10,
  plot.yTickLblSize = 10,
  plot.legendPosition = "right",
  plot.legendSize = 9,
  plot.Width = 170,
  plot.Height = 150
)
}
\value{
\code{ggplot2} graph object.
}
\description{

}
