% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileprocessing.R
\name{singleIMP}
\alias{singleIMP}
\title{singleIMP}
\usage{
singleIMP(tgtVar, methd = "mean")
}
\arguments{
\item{tgtVar}{Input single target variable.}

\item{methd}{Imputation method. Options are \code{"mean"}, \code{"random"} and \code{"regression"}. Default is \code{"mean"}.}
}
\value{
Outputs a \code{vector} object with all the missing value imputed.
}
\description{
data imputation for a vector of values.
}
\examples{
\dontrun{
mydfm <- singleIMP(myVector, methd = "mean")
}
}
