% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioUtil.R
\name{substr_right}
\alias{substr_right}
\title{substr_right}
\usage{
substr_right(x, n)
}
\arguments{
\item{x}{str. Input string.}

\item{n}{int. Optional user defined export name prefix. Default is \code{NULL}.}
}
\value{
character extracted from string.
}
\description{
Extract n'th character from string from right.
}
\examples{
\dontrun{
  substr_right("abc", 2)
}
}
