% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioUtil.R
\name{sv_importance_shapviz}
\alias{sv_importance_shapviz}
\title{sv_importance_shapviz}
\usage{
sv_importance_shapviz(
  object,
  kind = c("bar", "beeswarm", "both", "no"),
  max_display = 15L,
  fill = "#fca50a",
  bar_width = 2/3,
  bee_width = 0.4,
  bee_adjust = 0.5,
  viridis_args = getOption("shapviz.viridis_args"),
  color_bar_title = "Feature value",
  show_numbers = FALSE,
  format_fun = format_max,
  number_size = 3.2,
  sort_features = TRUE,
  ...
)
}
\value{
\code{ggplot2} graph object.
}
\description{
A alternative version of \code{sv_importance} function from \code{shapviz} package
}
\details{
Mainly to add custom graphic settings to the original function
}
