% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_svm_assessment.R
\name{svm_cv_rocauc_helper}
\alias{svm_cv_rocauc_helper}
\title{svm_cv_rocauc_helper}
\usage{
svm_cv_rocauc_helper(object, roc.smooth = FALSE, verbose = TRUE)
}
\arguments{
\item{object}{A \code{rbiosvm_cv} or \code{rbiosvm_nestedcv} object.}

\item{roc.smooth}{If to smooth the curves. Uses binomial method to smooth the curves. Default is \code{FALSE}.}

\item{verbose}{whether to display messages. Default is \code{TRUE}. This will not affect error or warning messages.}

\item{fileprefix}{String. A file prefix to use for export file name, instead of the object name. Default is \code{NULL}.}
}
\value{
Returns a \code{cv_auc_res} (also a \code{list}) class object for follow up use.
}
\description{
ROC-AUC helper function for cross-validated SVM models
}
\details{
Uses pROC module to calculate ROC. The function supports more than two groups or more than one threshold
         for classification and regression model.

         The test data included in the input \code{rbiosvm_nestedcv} is already normalized with training data information.
         So there is no need for additional data transformation.

         Also: a ROC curve with AUC == 1 is always 1-1 for 95% CI and can be misleading.

         The \code{cv_auc_res} class includes the following items:
         \code{cv_res_list}: a list containing CV ROC-AUC results organized per CV folds
         \code{smooth_roc}: if the ROC calculation is smoothed
         \code{input_model_class}: input CV model type, \code{rbiosvm_cv} (CV) or \code{rbiosvm_nestedcv} (nested CV)
}
\examples{
\dontrun{
rbioClass_svm_cv_roc_auc(object = svm_nested_cv)
}
}
