% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioUtil.R
\name{uni_PreProc}
\alias{uni_PreProc}
\title{uni_PreProc}
\usage{
uni_PreProc(rawlist, logTrans = FALSE, logTransMethod = "log2",
  logTransObjT = "data", logTransParallelComputing = FALSE,
  bgMethod = "auto", normMethod = "quantile", ...)
}
\arguments{
\item{rawlist}{Input data, either a list, \code{EList} or \code{MAList} object.}

\item{logTrans}{If to perfom a log transformation on the data or not. Default is \code{FALSE}.}

\item{logTransMethod}{If \code{logTrans = TRUE}, set which method to use for the transformation, \code{"log2"} or \code{"log10"}. Default is \code{"log2"}.}

\item{logTransObjT}{If \code{logTrans = TRUE}, set the file name for the output \code{csv} file containing the log transformed data.}

\item{logTransParallelComputing}{If \code{logTrans = TRUE}, set if to use parallel computing for the transformation or not. Default is \code{FALSE}.}

\item{bgMethod}{Background correction method. Default is \code{"auto"}. See \code{backgroundCorrect()} function from \code{limma} package for details.}

\item{normMethod}{Normalization method. Default is \code{"quantile"}. See \code{normalizeBetweenArrays()} function from \code{limma} package for details.}

\item{...}{arguments for \code{backgroundCorrect.matrix()} or \code{backgroundCorrect()} functions from \code{limma} package.}
}
\value{
Depending on the input type, the function outputs a \code{list}, \code{Elist} or \code{MAList} object with corrected and normalized expression values.
        If \code{logTrans = TRUE}, the function also outputs a \code{csv} file containing the log transformed data.
}
\description{
a legacy function borrowed from RBioArray. Data pre-processing function for the microarary data, which is now used for univariate reduction during svm nested CV.
}
\details{
The function does not use design matrix for array weight calculation.
         Therefore, the DE analysis based on the output from this function will yeild slightly different resutls from the \code{\link{rbioarray_filter_combine}}.
}
