% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_svm.R
\name{rbioClass_svm}
\alias{rbioClass_svm}
\title{rbioClass_svm}
\usage{
rbioClass_svm(
  x,
  y,
  center.scale = TRUE,
  kernel = c("radial", "linear", "polynomial", "sigmoid"),
  svm.cross.k = 10,
  tune.method = c("cross", "boot", "fix"),
  tune.cross.k = 10,
  tune.boot.n = 10,
  ...,
  n_cores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{Input data matrix (e.g., independent variables, predictors, features, X, etc). Make sure it is either a matrix or a dataframe.}

\item{y}{Input response variable (e.g.,dependent variables, Y etc). Make sure it is \code{factor} class.}

\item{center.scale}{Logical, whether center and scale the data, i.e. subtracting mean (col.mean) and dividing by standard deviation (col.sd). Default is \code{TRUE}.}

\item{kernel}{SVM kernel. Options are \code{"linear", "ploynomial", "radial", "sigmoid"}. Default is \code{"radial"}, aka RBF.}

\item{svm.cross.k}{Fold of cross validation. Default is \code{10}.}

\item{tune.method}{Parameter tuning method. Options are \code{"cross"} (i.e. cross validation), \code{"boot"} (i.e. bootstrap), and \code{"fix"}. Default is \code{"cross"}.}

\item{tune.cross.k}{Set only when \code{tune.method = "cross"}, fold number for cross validation. Default is \code{10}.}

\item{tune.boot.n}{Set only when \code{tune.method = "boot"}, bootstrap iterations. Default is \code{10}.}

\item{...}{Additional arguments for \code{svm} function from \code{e1071mc} package.}

\item{n_cores}{Number of CPU cores to use. Default is \code{NULL}, or the total number cores minus one.}

\item{verbose}{Whether to display messages. Default is \code{TRUE}. This will not affect error or warning messages.}
}
\value{
Returns a SVM model object, with classes "svm" and "rbiosvm".

Additional items for \code{rbiosvm} object to \code{svm} object from e1071mc package:

\code{inputX}: raw input predictor data.

\code{inputY}: input group labels.

\code{center.scaledX}: centered X data with scaling if applicable.

\code{class.weight}: class weight, \code{1} if data is balanced.

\code{svm.cross.k}: fold number for cross validation

\code{tune.method}: grid search method

\code{tune.cross.k}: fold number for grid search when using cross validation method, i.e. \code{tune.method = "cross"}.

\code{tune.boot.n}: n number for grid search when using bootstrap, i.e. \code{tune.method = "boot"}.

\code{model.type}: the SVM model type, "classification" or "regression".
}
\description{
Support Vector Machine (SVM) modelling
}
\details{
Model is trained with probability calculation enabled, so that \code{\link{rbioClass_svm_predict}} will be able calculate prediction probabilities.

Parameter tuning is for gamma (not applicable when \code{kernel = "linear"}) and cost.

The function automatically detects if the data is unbalanced or not, and applies class weight accordingly. For unbalanced data, the
class weight the inverse of the draw probability: \code{sample size / class size}.

When \code{tune.method = "cross"} and the sample size set for \code{tune.cross.k}, the method is effectively "leave-one-out (LOO)" cross-validation.
Same goes for \code{svm.cross.k} argument.

If \code{svm.cross.k} is set to 0, no cross validation is used for the final modelling.

The center.scale is \code{(x -  col.mean)/col.sd}. Raw data for SVM should be center.scaled. The training data column mean and column sd are used by the
center.scale process for the test data for ROC-AUC analysis. Usually, we center.scale the training data and use the col.mean and col.sd for center.scaling the test data.

The option \code{center.scale = FALSE} is for prior center.scaled whole data (training + test sets), i.e center scale the data then split for training and test sets.
Such process is only used for pure mathematical analysis of the data set, instead of the generalized use for SVM model evaluation and utility (i.e. classify unknown data).

When \code{n_cores>1}, parallel computing uses \code{e1071mc::svm_mc()} and \code{e1071mc::tune_mc()}. \code{n_cores=NULL} is the same as \code{n_cores=1}, which falls back to \code{e1071::svm()} and \code{e1017::tune()}

The function also supports regression study, in which case, the performance metric is \code{RMSE}.
}
\examples{
\dontrun{
svm_model <- rbioClass_svm(x = training_set[, -1], y = training_set[, 1], kernel = "radial", center = TRUE, scale = FALSE)
}
}
