% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_svm.R
\name{rbioClass_svm_ncv_fs}
\alias{rbioClass_svm_ncv_fs}
\title{rbioClass_svm_ncv_fs}
\usage{
rbioClass_svm_ncv_fs(x, y, center.scale = TRUE, kernel = "radial",
  cross.k = 10, tune.method = "cross", tune.cross.k = 10,
  tune.boot.n = 10, ..., fs.method = "rf", fs.count.cutoff = cross.k,
  parallelComputing = TRUE, clusterType = "PSOCK", verbose = TRUE)
}
\arguments{
\item{x}{Input data matrix (e.g., independent variables, predictors, features, X, etc). Make sure it is either a matrix or a dataframe.}

\item{y}{Input response variable (e.g.,dependent variables, Y etc). Make sure it is \code{factor} class.}

\item{center.scale}{Logical, wether center and scale the data, i.e. subtracting mean (col.mean) and deviding by standard deviation (col.sd). Default is \code{TRUE}.}

\item{kernel}{SVM kernel. Options are \code{"linear", "ploynomial", "radial", "sigmoid"}. Default is \code{"radial"}, aka RBF.}

\item{cross.k}{Fold of nested cross validation, i.e. outer loop. Default is \code{10}.}

\item{tune.method}{Parameter tuning method, i.e. innter loop. Options are \code{"cross"} (i.e. cross validation), \code{"boot"} (i.e. bootstrap), and \code{"fix"}. Default is \code{"cross"}.}

\item{tune.cross.k}{Set only when \code{tune.method = "cross"}, fold number for cross validation. Default is \code{10}.}

\item{tune.boot.n}{Set only when \code{tune.method = "boot"}, bootstrap iterations. Default is \code{10}.}

\item{...}{Additional arguments for \code{rbioClass_svm}.}

\item{fs.method}{Feature selection method. Only \code{"rf"} (i.e. random forest) is supported so far. Default is \code{"rf"}.}

\item{fs.count.cutoff}{A integer for feature vote cutoff. Default is outer loop cross-valiation fold \code{cross.k}.}

\item{parallelComputing}{Wether to use parallel computing or not. Default is \code{TRUE}.}

\item{clusterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}
}
\value{
Returns a SVM model object, with classes "svm" and "rbiosvm".

Additional items for \code{rbiosvm_nestedcv}:

\code{cv.fold}:

\code{randomized.sample.index}:

\code{tot.nested.accuracy.summary}:

\code{nested.accuracy}:

\code{fs.method}:

\code{selected.features}:

\code{nested.fs.count}:

\code{tune.method}:

\code{tune.cross.k}:

\code{tune.boot.n}:
}
\description{
Nested cross-validation assessment for SVM classification. It evaluates the overall performace of SVM modelling given the training data.
}
\details{
TBA

For now, RBioFS implementation of two-step random forest feature selection is used to select features based on nested cross-validation.
Resulted features from each nested cross-validation round are voted. Features with votes equal or greater than the cutoff are reported as selected features.
It is also a good idea to set \code{fs.count.cutoff} as \code{cross.k - 1}. Notably, \code{fs.count.cutoff = 1} is "no threshold",
meaning maximum number of features selected from the nested cross-validation are reported.
}
\examples{
\dontrun{
svm_nestedcv <- rbioClass_svm_ncv_fs(x = mydata[, -c(1:2)],
                                     y = factor(mydata$Conditions, levels = unique(mydata$Conditions)),
                                     center.scale = TRUE,
                                     cross.k = 5,
                                     tune.method = "cross",
                                     tune.cross.k = 5,
                                     parallelComputing = TRUE, clusterType = "FORK")

}
}
