% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_svm.R
\name{rbioClass_svm_ncv_fs}
\alias{rbioClass_svm_ncv_fs}
\title{rbioClass_svm_ncv_fs}
\usage{
rbioClass_svm_ncv_fs(
  x,
  y,
  univariate.fs = FALSE,
  uni.log2trans = TRUE,
  uni.contrast = NULL,
  uni.alpha = 0.05,
  uni.fdr = FALSE,
  center.scale = TRUE,
  kernel = c("radial", "linear", "polynomial", "sigmoid"),
  cross.k = 10,
  cross.best.model.method = c("none", "median"),
  tune.method = c("cross", "boot", "fix"),
  tune.cross.k = 10,
  tune.boot.n = 10,
  ...,
  fs.method = "rf",
  rf.ifs.ntree = 1001,
  rf.sfs.ntree = 1001,
  fs.count.cutoff = cross.k,
  parallelComputing = TRUE,
  n_cores = parallel::detectCores() - 1,
  clusterType = c("PSOCK", "FORK"),
  verbose = TRUE
)
}
\arguments{
\item{x}{Input data matrix (e.g., independent variables, predictors, features, X, etc). Make sure it is either a matrix or a dataframe.}

\item{y}{Input response variable (e.g.,dependent variables, Y etc). Make sure it is \code{factor} class.}

\item{univariate.fs}{If to use limma-based univariate reduction. Default is \code{FALSE}.}

\item{uni.log2trans}{Only set if \code{univariate.fs = TRUE}, if to log2 transform data before univariate reduction.}

\item{uni.contrast}{Only set if \code{univariate.fs = TRUE} and for a classification study, the contrast for the univariate analysis. Default is \code{NULL}.}

\item{uni.alpha}{Only set if \code{univariate.fs = TRUE}, the p value alpha. Default is \code{0.05}.}

\item{uni.fdr}{Only set if \code{univariate.fs = TRUE}, if to use FDR for the p value. Default is \code{FALSE}.}

\item{center.scale}{Logical, whether center and scale the data, i.e. subtracting mean (col.mean) and dividing by standard deviation (col.sd). Default is \code{TRUE}.}

\item{kernel}{SVM kernel. Options are \code{"linear", "ploynomial", "radial", "sigmoid"}. Default is \code{"radial"}, aka RBF.}

\item{cross.k}{Fold of nested cross validation, i.e. outer loop. Default is \code{10}.}

\item{cross.best.model.method}{The method to select the best cv models for feature selection. Options are \code{"median"} and \code{"none"}. Default is \code{"median"}.}

\item{tune.method}{Parameter tuning method, i.e. inner loop. Options are \code{"cross"} (i.e. cross validation), \code{"boot"} (i.e. bootstrap), and \code{"fix"}. Default is \code{"cross"}.}

\item{tune.cross.k}{Set only when \code{tune.method = "cross"}, fold number for cross validation. Default is \code{10}.}

\item{tune.boot.n}{Set only when \code{tune.method = "boot"}, bootstrap iterations. Default is \code{10}.}

\item{...}{Additional arguments for \code{rbioClass_svm}.}

\item{fs.method}{Feature selection method. Only \code{"rf"} (i.e. random forest) is supported so far. Default is \code{"rf"}.}

\item{rf.ifs.ntree}{Set only when \code{fs.method = "rf"}, ntree setting for the initial feature selection step of recursive feature selection. Default is \code{501}.}

\item{rf.sfs.ntree}{Set only when \code{fs.method = "rf"}, ntree setting for the sequential forward selection step of recursive feature selection. Default is \code{501}.}

\item{fs.count.cutoff}{A integer for feature vote cutoff. Default is outer loop cross-validation fold \code{cross.k}.}

\item{parallelComputing}{Whether to use parallel computing or not. Default is \code{TRUE}.}

\item{n_cores}{Only set when \code{parallelComputing = TRUE}, the number of CPU cores to use. Default is \code{detectCores() - 1}, or the total number cores minus one.}

\item{clusterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}

\item{verbose}{Whether to display messages. Default is \code{TRUE}. This will not affect error or warning messages.}
}
\value{
Returns a nested CV SVM model object: \code{rbiosvm_nestedcv}.

Additional items for \code{rbiosvm_nestedcv}:

\code{univariate.fs}

\code{cv.fold}: number of (outer) cross-validation fold

\code{randomized.sample.index}: randomized sample order for (outer) cross-validation. \code{NULL} for classification modelling (cv.fold already randomized).

\code{model.type}: the SVM model type, "classification" or "regression".

\code{tot.nested.accuracy.summary}: total (i.e. mean) nested cross-validation accuracy, if \code{model_type = "classification"}

\code{tot.nested.RMSE.summary}: total (i.e. mean) nested cross-validation RMSE, if \code{model_type = "regression"}

\code{tot.nested.rsq.summary}: total (i.e. mean) nested cross-validation R2, if \code{model_type = "regression"}

\code{nested.accuracy}: accuracy for each cross-validation iteration, if \code{model_type = "classification"}

\code{nested.RMSE}: RMSE for each cross-validation iteration, if \code{model_type = "regression"}

\code{nested.rsq}: R2 for each cross-validation iteration, if \code{model_type = "regression"}

\code{nested.fs.count}: fs count for all the cv models

\code{nested.cv.models}: all the cv models with test data. NOTE: the test data is center-scaled with training data if \code{center.scale = TRUE}.

\code{best.nested.method}: method to determine the best nested model.

\code{best.nested.index}: index of the best models in the total model list

\code{best.nested.acc.summary}

\code{best.nested.rmse.summary}

\code{best.nested.rsq.summary}

\code{best.nested.accu}

\code{best.nested.rmse}

\code{best.nested.rsq}

\code{best.nested.fs.count}: fs count for only the best models

\code{fs.method}: feature selection method

\code{fs.count.threshold}: count threshold for generating consensus feature list

\code{selected.features}

\code{tune.method}: (inner) loop method for SVM grid search

\code{tune.cross.k}: fold number for (inner) loop if cross-validation is chosen

\code{tune.boot.n}: iternation number for (inner) loop if bootstrap is chosen

\code{run.time}: total run time
}
\description{
Nested cross-validation feature selection and SVM modelling: CV-rRF-FS-SVM.
}
\details{
For now, RBioFS implementation of two-step random forest feature selection is used to select features based on nested cross-validation.
Resulted features from each nested cross-validation round are voted. Features with votes equal or greater than the cutoff are reported as selected features.

It is also a good idea to set \code{fs.count.cutoff} as \code{cross.k - 1}. Notably, \code{fs.count.cutoff = 1} is "no threshold",
meaning maximum number of features selected from the nested cross-validation are reported.

When \code{cross.best.model.method = "median"}, the function only use models with accuracy/RMSE equal or better than the median value
for feature count threholding. When there is no change in performance across cv models, the function behaves same as \code{cross.best.model.method = "none"}


The function also supports regression study, in which case, the performance metric is \code{RMSE}.

For classification modelling, the 10-fold CV segmentation is stratified, whereas the "normal" random resampling is used for regression modelling.
}
\examples{
\dontrun{
svm_nestedcv <- rbioClass_svm_ncv_fs(x = mydata[, -c(1:2)],
                                     y = factor(mydata$Conditions, levels = unique(mydata$Conditions)),
                                     center.scale = TRUE,
                                     cross.k = 5,
                                     tune.method = "cross",
                                     tune.cross.k = 5,
                                     fs.method = "rf",
                                     rf.ifs.ntree = 1001, rf.sfs.ntree = 1001,
                                     parallelComputing = TRUE, clusterType = "FORK")

}
}
