% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_svm.R
\name{rbioClass_svm_roc_auc}
\alias{rbioClass_svm_roc_auc}
\title{rbioClass_svm_roc_auc}
\usage{
rbioClass_svm_roc_auc(
  object,
  newdata = NULL,
  newdata.label = NULL,
  center.scale.newdata = TRUE,
  rocplot = TRUE,
  plot.smooth = FALSE,
  plot.SymbolSize = 2,
  plot.lineSize = 1,
  plot.display.Title = TRUE,
  plot.titleSize = 10,
  plot.fontType = "sans",
  plot.xLabel = "1 - specificity",
  plot.xLabelSize = 10,
  plot.xTickLblSize = 10,
  plot.yLabel = "sensitivity",
  plot.yLabelSize = 10,
  plot.yTickLblSize = 10,
  plot.legendSize = 9,
  plot.rightsideY = TRUE,
  plot.Width = 170,
  plot.Height = 150,
  verbose = TRUE
)
}
\arguments{
\item{object}{A \code{rbiosvm} object.}

\item{newdata}{A data matrix or vector for test data. Make sure it is a \code{matrix} or \code{vector} without labels, as well as the same feature numbers as the training set.}

\item{newdata.label}{The correspoding label vector to the data. Make sure it is a \code{factor} object. Defaults is \code{NULL}.}

\item{center.scale.newdata}{Logical, wether center and scale the newdata with training data mean and standard deviation. Default is \code{TRUE}.}

\item{rocplot}{If to generate a ROC plot. Default is \code{TRUE}.}

\item{plot.smooth}{If to smooth the curves. Uses binormal method to smooth the curves. Default is \code{FALSE}.}

\item{plot.SymbolSize}{Symbol size. Default is \code{2}.}

\item{plot.lineSize}{Line size. Default is \code{1}.}

\item{plot.display.Title}{If to show the name of the y class. Default is \code{TRUE}.}

\item{plot.titleSize}{The font size of the plot title. Default is \code{10}.}

\item{plot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is avaiable on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{plot.xLabel}{X-axis label. Type with quotation marks. Could be NULL. Default is \code{"1 - specificity"}.}

\item{plot.xLabelSize}{X-axis label size. Default is \code{10}.}

\item{plot.xTickLblSize}{X-axis tick label size. Default is \code{10}.}

\item{plot.yLabel}{Y-axis label. Type with quotation marks. Could be NULL. Default is \code{"sensitivity"}.}

\item{plot.yLabelSize}{Y-axis label size. Default is \code{10}.}

\item{plot.yTickLblSize}{Y-axis tick label size. Default is \code{10}.}

\item{plot.legendSize}{Legend size. Default is \code{9}.}

\item{plot.rightsideY}{If to show the right side y-axis. Default is \code{FALSE}.}

\item{plot.Width}{Scoreplot width. Default is \code{170}.}

\item{plot.Height}{Scoreplot height. Default is \code{150}.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}

\item{plot.comps}{Number of comps to plot. Default is \code{1:object$ncomp}}
}
\value{
Prints AUC values in the console. And a pdf file for ROC plot. The function also exports a ROC results list as a \code{svm_roc_auc} class to the environment.

        Items of the \code{svm_roc_auc} class:
        \code{model.type}
        \code{y.threshold}
        \code{regression.categories}
        \code{svm.roc_object}
        \code{svm.roc_dataframe}
        \code{input.newdata}
        \code{input.newdata.y}
        \code{input.newdata.label}
        \code{newdata.center.scaled}
}
\description{
ROC-AUC analysis and ploting for SVM model, for classification model only.
}
\details{
Uses pROC module to calculate ROC. The function supports more than two groups or more than one threshold for classification and regression model.

         When \code{newdata} is not provided, the function uses the training data from the input SVM object.

         Although optional, the \code{newdata} matrix should use training data's column mean and column standard deviation to center.scale prior to ROC-AUC analysis.
         The option \code{center.scaled.newdata = FALSE} is used when the whole (training and test sets) data were center.scaled before SVM training and testing.

         For regression models, it is ok to provide multiple thresholds for \code{y.threshold}.

         Along with the data frame, the \code{svm_roc_auc} output includes the \code{roc} object from the \code{pROC} package, with which the stats can be done to compare ROCs.

         Also: a ROC curve with AUC == 1 is always 1-1 for 95% CI and can be misleading.
}
\examples{
\dontrun{
rbioClass_svm_roc_auc(object = svm_m, newdata = svm_test[, -1],
                      newdata.label = factor(svm_test$y, levels = unique(svm_test$y)))
}
}
