% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_svm.R
\name{rbioClass_svm_shap_individual}
\alias{rbioClass_svm_shap_individual}
\title{rbioClass_svm_shap_individual}
\usage{
rbioClass_svm_shap_individual(
  model,
  X,
  bg_X,
  bg_n = 200L,
  y = NULL,
  parallelComputing = FALSE,
  n_cores = parallel::detectCores() - 1,
  clusterType = c("PSOCK", "FORK"),
  plot.filename.prefix = NULL,
  plot.type = c("waterfall", "force"),
  plot.n = 15L,
  ...,
  plot.SymbolSize = 2,
  plot.lineSize = 1,
  plot.display.Title = TRUE,
  plot.titleSize = 10,
  plot.fontType = "sans",
  plot.bee.colorscale = "A",
  plot.bar.color = "blue",
  plot.xLabel = "SHAP value",
  plot.xLabelSize = 10,
  plot.xTickLblSize = 10,
  plot.yLabel = "Features",
  plot.yLabelSize = 10,
  plot.yTickLblSize = 10,
  plot.legendPosition = "right",
  plot.legendSize = 9,
  plot.Width = 15,
  plot.Height = 10,
  verbose = TRUE
)
}
\value{
The function outputs a \code{rbio_shap} class object.
}
\description{
Individualized SHAP analysis for SVM model.
}
\details{
1. This fucntion relies on \code{kernelshap} and \code{shapviz} logics.
   2. (To be tested) The function should work with regression SVM models as well.
}
