% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_svm_assessment.R
\name{rbioClass_svm_shap_individual}
\alias{rbioClass_svm_shap_individual}
\title{rbioClass_svm_shap_individual}
\usage{
rbioClass_svm_shap_individual(
  model,
  X,
  bg_X,
  bg_n = 200L,
  y = NULL,
  parallelComputing = FALSE,
  n_cores = parallel::detectCores() - 1,
  clusterType = c("PSOCK", "FORK"),
  plot.filename.prefix = NULL,
  plot.type = c("waterfall", "force"),
  plot.n = 15L,
  ...,
  plot.SymbolSize = 2,
  plot.lineSize = 1,
  plot.display.Title = TRUE,
  plot.titleSize = 10,
  plot.fontType = "sans",
  plot.bee.colorscale = "A",
  plot.bar.color = "blue",
  plot.xLabel = "Prediction value",
  plot.xLabelSize = 10,
  plot.xTickLblSize = 10,
  plot.yLabel = "Features",
  plot.yLabelSize = 10,
  plot.yTickLblSize = 10,
  plot.legendPosition = "right",
  plot.legendSize = 9,
  plot.Width = 170,
  plot.Height = 150,
  verbose = TRUE
)
}
\arguments{
\item{model}{Input SVM model object and should be a \code{rbiosvm} class.}

\item{X}{Input data, matrix or  data.frame, with column: features, and row: records. It should not contain outcome column.}

\item{bg_X}{Background data, \code{matrix} or \code{data.frame}, column: features, and row: records. It should not contain outcome column.}

\item{bg_n}{Number of sample to draw from the background data.}

\item{parallelComputing}{whether to use parallel computing or not. Default is \code{TRUE}.}

\item{n_cores}{Only set when \code{parallelComputing = TRUE}, the number of CPU cores to use. Default is \code{detectCores() - 1}, or the total number cores minus one.}

\item{clusterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}

\item{plot.filename.prefix}{Prefix string to add to the output plot file.}

\item{plot.type}{The type of the individual SHAP plot, options are \code{"waterfall", "force"}.}

\item{plot.n}{Number of features to display in the SHAP plot.}

\item{...}{Additional arguments for the \code{shapviz::sv_waterfall} or \code{shapviz::sv_force} functions.}

\item{plot.SymbolSize}{Symbol size. Default is \code{2}.}

\item{plot.lineSize}{Line size. Default is \code{1}.}

\item{plot.display.Title}{If to show the name of the y class. Default is \code{TRUE}.}

\item{plot.titleSize}{The font size of the plot title. Default is \code{10}.}

\item{plot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is avaiable on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{plot.xLabel}{X-axis label. Type with quotation marks. Could be NULL. Default is \code{"1 - specificity"}.}

\item{plot.xLabelSize}{X-axis label size. Default is \code{10}.}

\item{plot.xTickLblSize}{X-axis tick label size. Default is \code{10}.}

\item{plot.yLabel}{Y-axis label. Type with quotation marks. Could be NULL. Default is \code{"sensitivity"}.}

\item{plot.yLabelSize}{Y-axis label size. Default is \code{10}.}

\item{plot.yTickLblSize}{Y-axis tick label size. Default is \code{10}.}

\item{plot.legendSize}{Legend size. Default is \code{9}.}

\item{plot.Width}{Plot width, in \code{mm} unit. Default is \code{15}}

\item{plot.Height}{Plot height, in \code{mm} unit.. Default is \code{10}.}

\item{verbose}{whether to display messages. Default is \code{TRUE}. This will not affect error or warning messages.}

\item{plot}{If to generate aggregated SHAP plot.}
}
\value{
The function outputs a \code{rbio_shap} class object.
}
\description{
Individualized SHAP analysis for SVM model.
}
\details{
1. This fucntion relies on \code{kernelshap} and \code{shapviz} logic.
   2. (To be tested) The function should work with regression SVM models as well.
}
