% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioFS.R
\name{rbioFS}
\alias{rbioFS}
\title{rbioFS}
\usage{
rbioFS(
  objTitle = "data",
  rf_type = c("classification", "regression"),
  file = NULL,
  input = NULL,
  sampleIDVar = NULL,
  groupIDVar = NULL,
  annotVarNames = c(sampleIDVar, groupIDVar),
  impute = FALSE,
  imputeMethod = "rf",
  imputeIter = 10,
  imputeNtree = 501,
  quantileNorm = FALSE,
  nTimes = 50,
  nTree = 1001,
  parallelComputing = TRUE,
  n_cores = parallel::detectCores() - 1,
  clusterType = "PSOCK",
  plot = TRUE,
  initialFS_n = "all",
  initialFS_Title = NULL,
  initialFS_xLabel = "Mean Decrease in Accuracy",
  initialFS_yLabel = NULL,
  initialFS_errorbar = "SEM",
  initialFS_errorbarWidth = 0.2,
  initialFS_xTickLblSize = 10,
  initialFS_yTickLblSize = 10,
  initialFS_plotWidth = 170,
  initialFS_plotHeight = 150,
  SFS_n = "all",
  SFS_Title = NULL,
  SFS_xLabel = NULL,
  SFS_yLabel = NULL,
  SFS_errorbar = "SEM",
  SFS_errorbarWidth = 0.2,
  SFS_symbolSize = 2,
  SFS_xTickLblSize = 10,
  SFS_yTickLblSize = 10,
  SFS_plotWidth = 170,
  SFS_plotHeight = 150,
  verbose = TRUE
)
}
\arguments{
\item{objTitle}{The title for the output data frame. Default is \code{"data"}}

\item{rf_type}{RF modelling type, either \code{"classification"} or \code{"regression"}. Default is \code{"classifiation"}.}

\item{file}{Input file. Only takes \code{.csv} format, with the label column, i.e. the whole dataset. Set either this or \code{input}, but not both.}

\item{input}{Input data frame. The type should be \code{data.frame} or \code{matrix}, with the label column, i.e. the whole dataset. Set either this or \code{file}, but not both.}

\item{sampleIDVar}{Sample variable name. It's a character string.}

\item{groupIDVar}{Group variable name. It's a character string.}

\item{annotVarNames}{String, identifying the annotation variable names to exclude from the FS analysis. Default is \code{c(sampleIDVar, groupIDVar)}.}

\item{impute}{Wether to use data imputation functionality to impute missing data points. Default is \code{FALSE}.}

\item{imputeMethod}{The method for data imputation, only set if \code{impute = TRUE}. Default is \code{"rf"}. See \code{\link{rbioIMP}} for details.}

\item{imputeIter}{RF imputation iteration value. Default is \code{10}. See \code{\link{rbioIMP}} for details.}

\item{imputeNtree}{RF imputation ntree value. Default is \code{501}. See \code{\link{rbioIMP}} for details.}

\item{quantileNorm}{Wehter ot use quantile nomalization on the raw data. Default is \code{FALSE}.}

\item{nTimes}{Number of random forest for both initial FS and SFS-like FS. Default is \code{50} times.}

\item{nTree}{Number of trees generated for each random forest run for both initial FS and SFS-like FS. Default is \code{1001} trees.}

\item{parallelComputing}{Wether to use parallel computing or not. Default is \code{TRUE}.}

\item{n_cores}{Only set when \code{parallelComputing = TRUE}, the number of CPU cores to use. Default is \code{detectCores() - 1}, or the total number cores minus one.}

\item{clusterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}

\item{plot}{If to plot results, bar graph for initial FS and joint-point curve for SFS-like FS. Default is \code{TRUE}}

\item{initialFS_n}{Number of features to show. Takes integer numbers. Default is \code{"all"} (make sure to include quotation marks).}

\item{initialFS_Title}{Figure title. Make sure to use quotation marks. Use \code{NULL} to hide. Default is \code{NULL}.}

\item{initialFS_xLabel}{X-axis label. Make sure to use quotation marks. Use \code{NULL} to hide. Default is \code{NULL}.}

\item{initialFS_yLabel}{Y-axis label. Make sure to use quotatio marks. Use \code{NULL} to hide. Default is \code{"Mean Decrease in Accurac"}}

\item{initialFS_errorbar}{The type of errorbar in the graph. Options are \code{"SEM"} (standard error of the mean) or \code{"SD"} (standard deviation). Default is \code{"SEM"}.}

\item{initialFS_errorbarWidth}{The width of the errorbar. Default is \code{0.2}.}

\item{initialFS_xTickLblSize}{Font size for the x-axis text. Default is \code{10}.}

\item{initialFS_yTickLblSize}{Font size for the y-axis text. Default is \code{10}.}

\item{initialFS_plotWidth}{The width of the figure for the final output figure file. Default is \code{170}.}

\item{initialFS_plotHeight}{The height of the figure for the final output figure file. Default is \code{150}.}

\item{SFS_n}{Number of features to show. Takes integer numbers. Default is \code{"all"} (make sure to include quotation marks).}

\item{SFS_xLabel}{X-axis label. Make sure to use quotation marks. Use \code{NULL} to hide. Default is \code{NULL}.}

\item{SFS_yLabel}{Y-axis label. Make sure to use quotatio marks. Use \code{NULL} to hide. Default is \code{"Mean Decrease in Accurac"}}

\item{SFS_errorbar}{The type of errorbar in the graph. Options are \code{"SEM"} (standard error of the mean) or \code{"SD"} (standard deviation). Default is \code{"SEM"}.}

\item{SFS_errorbarWidth}{The width of the errorbar. Default is \code{0.2}.}

\item{SFS_symbolSize}{Size of the symbol. Default is \code{2}.}

\item{SFS_xTickLblSize}{Font size for the x-axis text. Default is \code{10}.}

\item{SFS_yTickLblSize}{Font size for the y-axis text. Default is \code{10}.}

\item{SFS_plotWidth}{The width of the figure for the final output figure file. Default is \code{170}.}

\item{SFS_plotHeight}{The height of the figure for the final output figure file. Default is \code{150}.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}
}
\value{
Outputs two \code{list} objects and \code{.csv} for the two FS steps. And, if \code{plot = TRUE}, a bargraph for initial FS and joint-point curve for SFS-like FS step. A \code{.csv} file with imputed data is also generated if \code{impute = TRIE}.
}
\description{
Recursive nested random forest variable importance (vi) and OOB error rate computation in a sequential forward selection (SFS) manner.
}
\details{
Make sure to arrange input \code{.csv} file with first two columns for smaple ID and conditions, and the rest for features (e.g., genes).
         \code{groupIDVar} also takes continuous variable for regression study.
}
\examples{
\dontrun{
rbioFS(file = "test.csv", impute = TRUE, imputeIter = 50, quantileNorm = TRUE)
}
}
