% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioFS.R
\name{rbioFS}
\alias{rbioFS}
\title{rbioFS}
\usage{
rbioFS(file, impute = FALSE, imputeMethod = "rf", imputeIter = 10,
  imputeNtree = 501, quantileNorm = FALSE, nTimes = 50, nTree = 1001,
  SFS_mTry = "recur_default", multicore = TRUE, plot = TRUE,
  initialFS_n = "all", initialFS_Title = NULL,
  initialFS_xLabel = "Mean Decrease in Accuracy", initialFS_yLabel = NULL,
  initialFS_errorbar = "SEM", initialFS_errorbarWidth = 0.2,
  initialFS_xTxtSize = 10, initialFS_yTxtSize = 10,
  initialFS_plotWidth = 170, initialFS_plotHeight = 150, SFS_n = "all",
  SFS_Title = NULL, SFS_xLabel = NULL, SFS_yLabel = NULL,
  SFS_errorbar = "SEM", SFS_errorbarWidth = 0.2, SFS_symbolSize = 2,
  SFS_xTxtSize = 10, SFS_yTxtSize = 10, SFS_plotWidth = 170,
  SFS_plotHeight = 150)
}
\arguments{
\item{file}{Input file. Only takes \code{.csv} format.}

\item{impute}{Wether to use data imputation functionality to impute missing data points. Default is \code{FALSE}.}

\item{imputeMethod}{The method for data imputation, only set if \code{impute = TRUE}. Default is \code{"rf"}. See \code{\link{rbioIMP}} for details.}

\item{imputeIter}{RF imputation iteration value. Default is \code{10}. See \code{\link{rbioIMP}} for details.}

\item{imputeNtree}{RF imputation ntree value. Default is \code{501}. See \code{\link{rbioIMP}} for details.}

\item{quantileNorm}{Wehter ot use quantile nomalization on the raw data. Default is \code{FALSE}.}

\item{nTimes}{Number of random forest for both initial FS and SFS-like FS. Default is \code{50} times.}

\item{nTree}{Number of trees generated for each random forest run for both initial FS and SFS-like FS. Default is \code{1001} trees.}

\item{SFS_mTry}{Number of randomly selected featurs for constructing trees for SFS-like FS step. When \code{"recur_default"}, it'll be based on \code{p / 3}; when \code{"rf_default"}, it will use the default setting in \code{randomForest} package. Default is \code{"recur_default"}.}

\item{multicore}{If to use parallel computing. Default is \code{TRUE}.}

\item{plot}{If to plot results, bar graph for initial FS and joint-point curve for SFS-like FS. Default is \code{TRUE}}

\item{initialFS_n}{Number of features to show. Takes integer numbers. Default is \code{"all"} (make sure to include quotation marks).}

\item{initialFS_Title}{Figure title. Make sure to use quotation marks. Use \code{NULL} to hide. Default is \code{NULL}.}

\item{initialFS_xLabel}{X-axis label. Make sure to use quotation marks. Use \code{NULL} to hide. Default is \code{NULL}.}

\item{initialFS_yLabel}{Y-axis label. Make sure to use quotatio marks. Use \code{NULL} to hide. Default is \code{"Mean Decrease in Accurac"}}

\item{initialFS_errorbar}{The type of errorbar in the graph. Options are \code{"SEM"} (standard error of the mean) or \code{"SD"} (standard deviation). Default is \code{"SEM"}.}

\item{initialFS_errorbarWidth}{The width of the errorbar. Default is \code{0.2}.}

\item{initialFS_xTxtSize}{Font size for the x-axis text. Default is \code{10}.}

\item{initialFS_yTxtSize}{Font size for the y-axis text. Default is \code{10}.}

\item{initialFS_plotWidth}{The width of the figure for the final output figure file. Default is \code{170}.}

\item{initialFS_plotHeight}{The height of the figure for the final output figure file. Default is \code{150}.}

\item{SFS_n}{Number of features to show. Takes integer numbers. Default is \code{"all"} (make sure to include quotation marks).}

\item{SFS_xLabel}{X-axis label. Make sure to use quotation marks. Use \code{NULL} to hide. Default is \code{NULL}.}

\item{SFS_yLabel}{Y-axis label. Make sure to use quotatio marks. Use \code{NULL} to hide. Default is \code{"Mean Decrease in Accurac"}}

\item{SFS_errorbar}{The type of errorbar in the graph. Options are \code{"SEM"} (standard error of the mean) or \code{"SD"} (standard deviation). Default is \code{"SEM"}.}

\item{SFS_errorbarWidth}{The width of the errorbar. Default is \code{0.2}.}

\item{SFS_symbolSize}{Size of the symbol. Default is \code{2}.}

\item{SFS_xTxtSize}{Font size for the x-axis text. Default is \code{10}.}

\item{SFS_yTxtSize}{Font size for the y-axis text. Default is \code{10}.}

\item{SFS_plotWidth}{The width of the figure for the final output figure file. Default is \code{170}.}

\item{SFS_plotHeight}{The height of the figure for the final output figure file. Default is \code{150}.}

\item{objTitle}{The title for the output data frame. Default is \code{"x_vs_tgt"}}
}
\value{
Outputs two \code{list} objects and \code{.csv} for the two FS steps. And, if \code{plot = TRUE}, a bargraph for initial FS and joint-point curve for SFS-like FS step. A \code{.csv} file with imputed data is also generated if \code{impute = TRIE}.
}
\description{
Recursive nested random froest variable importance (vi) and OOB error rate computation in a sequential forward selection (SFS) manner.
}
\details{
Make sure to arrange input \code{.csv} file with first two columns for smaple ID and conditions, and the rest for features (e.g., genes).
}
\examples{
\dontrun{
rbioFS(file = "test.csv", impute = TRUE, imputeIter = 50, quantileNorm = TRUE)
}
}

