% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioFS_PCA.R
\name{rbioFS_PCA}
\alias{rbioFS_PCA}
\alias{rbioFS_PCA.default}
\title{rbioFS_PCA}
\usage{
rbioFS_PCA(...)

\method{rbioFS_PCA}{default}(objTitle = "data", input, idx,
  scaleData = TRUE, boxplotTitle = NULL, boxplotWidth = 170,
  boxplotHeight = 150, biplotPC = c("PC1", "PC2"), biplotTitle = NULL,
  biplotSymbolSize = 2, ellipse = FALSE, ellipse_conf = 0.93,
  loadingPlot = TRUE, loadingSize = 3, biplotWidth = 170,
  biplotHeight = 150, fontType = "sans", xTickLblSize = 10,
  yTickLblSize = 10)
}
\arguments{
\item{objTitle}{File name prefix for output files and the PCA object. Default is \code{"data"}.}

\item{input}{Input data, data frame.}

\item{idx}{Sample index. It is a factor object.}

\item{scaleData}{If to scale the data when performing PCA. Default is \code{TRUE}.}

\item{boxplotTitle}{The boxplot title. Default is \code{NULL}.}

\item{boxplotWidth}{The boxplot width. Default is \code{170}.}

\item{boxplotHeight}{The boxplot height. Default is \code{150}.}

\item{biplotPC}{The two PCs to plot for the biplot. Default is \code{c("PC1", "PC2")}.}

\item{biplotTitle}{The biplot title. Default is \code{NULL}.}

\item{biplotSymbolSize}{The symbol size for the scatter plot portion of the biplot. Default is \code{2}.}

\item{ellipse}{If to draw ellipses. Default is \code{FALSE}.}

\item{ellipse_conf}{The confidence value for the ellipses. Default is \code{0.93}.}

\item{loadingPlot}{If to superimpose loading plot. Default is \code{TRUE}.}

\item{loadingSize}{The font size of the loading plot labels. Default is \code{3}.}

\item{biplotWidth}{The biplot width. Default is \code{170}.}

\item{biplotHeight}{The biplot height. Default is \code{150}.}

\item{fontType}{Font for the figure texts.}

\item{file}{Input file name. Format is \code{csv}.}

\item{xLabel_biplot}{X-axis label for the biplot. Default is NULL.}

\item{yLabel_biplot}{Y-axis label for the biplot. Default is NULL.}
}
\value{
Outputs a PCA object, a boxplot (proportion of variance) and a biplot from PCA analysis. The format is \code{pdf}.
}
\description{
A simple to use wrapper for PCA (Principal Component Analysis) and visualization
}
\details{
Make sure to arrange input data with first two columns for smaple ID and conditions, and the rest for features (e.g., genes).
}
\examples{
\dontrun{
rbioFS_PCA(input = pcaDfm, idx = pcaDfm$Conditions, ellipse = TRUE, loadingPlot = TRUE, biplotWidth = 200, biplotHeight = 170)
}
}
