% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioFS_plsda.R
\name{rbioFS_plsda_perm}
\alias{rbioFS_plsda_perm}
\title{rbioFS_plsda_perm()}
\usage{
rbioFS_plsda_perm(object, nperm = 999, ncomp = object$ncomp,
  adjCV = FALSE, parallelComputing = TRUE, clusterType = "PSOCK",
  verbose = TRUE)
}
\arguments{
\item{object}{A \code{rbiomvr} or \code{mvr} object. Make sure the object is generated with a \code{validation} section.}

\item{nperm}{Number of permutations to run. Default is \code{999}.}

\item{ncomp}{tbc}

\item{adjCV}{tbc}

\item{parallelComputing}{Wether to use parallel computing or not. Default is \code{TRUE}.}

\item{clusterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}.}
}
\value{
tbc
}
\description{
Permutation test for PLS-DA models.
}
\details{
tbc
}
\examples{
\dontrun{
rbioFS_plsda_perm()
}
}
