% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioFS_plsda.R
\name{rbioFS_plsda_predict}
\alias{rbioFS_plsda_predict}
\title{rbioFS_plsda_predict}
\usage{
rbioFS_plsda_predict(object, comps = object$ncomp, newdata,
  prob.method = "Bayes", threshold = 0.2, predplot = TRUE,
  plot.sampleLabel.type = "none", plot.sampleLabel.vector = NULL,
  plot.sampleLabelSize = 2, plot.sampleLabel.padding = 0.5,
  multi_plot.ncol = length(levels(object$inputY)), multi_plot.nrow = 1,
  multi_plot.legend.pos = "bottom", plot.rightsideY = TRUE,
  plot.SymbolSize = 2, plot.display.Title = TRUE, plot.titleSize = 10,
  plot.fontType = "sans", plot.unclassifiedColour = "gray",
  plot.classifiedColour = "red", plot.xLabel = "Samples",
  plot.xLabelSize = 10, plot.xTickLblSize = 6, plot.xTickItalic = FALSE,
  plot.xAngle = 0, plot.xhAlign = 0.5, plot.xvAlign = 0.5,
  plot.yLabel = "Predicted values", plot.yLabelSize = 10,
  plot.yTickLblSize = 10, plot.legendSize = 9, plot.Width = 170,
  plot.Height = 150)
}
\arguments{
\item{object}{A \code{rbiomvr} or \code{mvr} object.}

\item{comps}{Number of PLS-DA components used in the model. Default is \code{object$ncomp}.}

\item{newdata}{Input data to be classified. Make sure it is a \code{matrix} class and has the same variables as the model, i.e. same number of columns as the training data.}

\item{prob.method}{Method to calculate classification probability. Options are \code{"softmax"} and \code{"Bayes"}. See details for more information. Default is \code{"Bayes"}.}

\item{threshold}{Classification threshold. Should be a number between \code{0} and \code{1}. Default is \code{0.2}.}

\item{predplot}{If to generate a prediction value plot. Default is \code{TRUE}.}

\item{plot.sampleLabel.type}{If to show the sample labels on the graph. Options are \code{"none"}, \code{"direct"} and \code{"indirect"}. Default is \code{"none"}.}

\item{plot.sampleLabel.vector}{Set only when \code{plot.sampleLabel.type} is not set to \code{"none"}, a character vector containing annotation (i.e. labels) for the samples. Default is \code{NULL}.}

\item{plot.sampleLabelSize}{Only set when \code{plot.sampleLabel.type} is not \code{"none"}. The size of the sample label. Default is \code{2}.}

\item{plot.sampleLabel.padding}{Set only when \code{plot.sampleLabel.type = "indirect"}, the padding between sample symbol and the label. Default is \code{0.5}.}

\item{multi_plot.ncol}{Number of columns on one figure page. Default is \code{length(plot.comps)}.}

\item{multi_plot.nrow}{Number of rows on one figure page. Default is \code{1}.}

\item{multi_plot.legend.pos}{The legend position. Only effective when multi-plot is generated. Options are \code{"bottom"}, \code{"top"}, \code{"left"} and \code{"right"}. Default is \code{"bottom"}.}

\item{plot.rightsideY}{If to show the right side y-axis. Default is \code{FALSE}. Note: doesn't seem to be necessasry as PLS-DA always has at least two y classes.}

\item{plot.SymbolSize}{Symbol size. Default is \code{2}.}

\item{plot.display.Title}{If to show the name of the y class. Default is \code{TRUE}.}

\item{plot.titleSize}{The font size of the plot title. Default is \code{10}.}

\item{plot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is avaiable on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{plot.unclassifiedColour}{Colour for the unclassified samples. Default is \code{"gray"}.}

\item{plot.classifiedColour}{Colour for the classified samples. Default is \code{"red"}.}

\item{plot.xLabel}{X-axis label. Type with quotation marks. Could be NULL. Default is \code{"Samples"}.}

\item{plot.xLabelSize}{X-axis label size. Default is \code{10}.}

\item{plot.xTickLblSize}{X-axis tick label size. Default is \code{10}.}

\item{plot.xAngle}{The rotation angle (degrees) of the x axis marks. Default is \code{0} - horizontal.}

\item{plot.xhAlign}{The horizontal alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{plot.xvAlign}{The vertical alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{plot.yLabel}{Y-axis label. Type with quotation marks. Could be NULL. Default is \code{"Predicted values"}.}

\item{plot.yLabelSize}{Y-axis label size. Default is \code{10}.}

\item{plot.yTickLblSize}{Y-axis tick label size. Default is \code{10}.}

\item{plot.legendSize}{Legend size. Default is \code{9}.}

\item{plot.Width}{Scoreplot width. Default is \code{170}.}

\item{plot.Height}{Scoreplot height. Default is \code{150}.}
}
\value{
A \code{prediction} obejct, as well as pdf figure file for predicted values if \code{predplot = TRUE}.
}
\description{
Prediction function for PLS-DA analysis. The function calculates the predicted value for unknown sample data using the input PLS-DA model.
}
\details{
Regarding \code{threshold}, the value should between \code{0} and \code{1}. It's the flank region around the dummified classification values \code{0} (i.e. "control") and \code{1} (i.e. "case").
This is only for information sake.

The "Bayes" method uses the klaR package implementation of naive Bayes algorithm, based on the Bayes theorem: \code{P(A|B) = P(B|A)P(A)/P(B)}.
Specifically, the equation models classification against predicted values from pls-da model and training data: \code{P(class|predicted values) = P(predicted values|)P(class)/P(predicted values)}.
The resulted Bayesian classification model is then applied to newdata, thereby prosterior probabilites are calcuated.

The "softmax" method uses the equation: \code{probability = exp(predicted.value) / sum(exp(predicted.values))}. This method doesn't relay on training data.

The conventional wisdom is to use "Bayes" method for unbalanced classification, and {"softmax"} for balanced situation.

For classification plot, the output \code{prediction} object should be used with function \code{\link{rbioFS_plsda_classplot()}}.
}
\examples{
\dontrun{
rbioFS_plsda_predict(object = new_model_optm, newdata = newdata, prob.method = "Bayes",
                     plot.sampleLabel.type = "none", plot.sampleLabel.vector = NULL, plot.sampleLabel.padding = 0.5,
                     multi_plot.ncol = length(levels(object$inputY)), multi_plot.nrow = 1, multi_plot.legend.pos = "bottom",
                     plot.SymbolSize = 2, plot.display.Title = TRUE, plot.titleSize = 10,
                     plot.fontType = "sans", plot.unclassifiedColour = "gray", plot.classifiedColour = "red",
                     plot.xLabel = "Samples", plot.xLabelSize = 10, plot.xTickLblSize = 6, plot.xTickItalic = FALSE,
                     plot.xAngle = 0, plot.xhAlign = 0.5, plot.xvAlign = 0.5,
                     plot.yLabel = "Predicted values", plot.yLabelSize = 10, plot.yTickLblSize = 10,
                     plot.legendSize = 9,
                     plot.Width = 170, plot.Height = 150)
}
}
