% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioClass_plsda.R
\name{rbioFS_plsda_vip}
\alias{rbioFS_plsda_vip}
\title{rbioFS_plsda_vip}
\usage{
rbioFS_plsda_vip(
  object,
  vip.alpha = 1,
  comps = c(1, 2),
  bootstrap = TRUE,
  boot.n = 50,
  boot.parallelComputing = TRUE,
  boot.n_cores = parallel::detectCores() - 1,
  boot.clusterType = "PSOCK",
  plot = TRUE,
  ...,
  verbose = TRUE
)
}
\arguments{
\item{object}{A \code{mvr} or \code{rbiomvr} object. Make sure the model is built uisng \code{"oscorespls"} method.}

\item{vip.alpha}{Alpha value (threshold) for VIP values. Any VIP above this is considered important. Defaults is \code{1}.}

\item{comps}{Integer vector. Components to plot. The index of the components are intergers. The vector length should be between 1 and the total number of components, inclusive. Default is \code{c(1, 2)}.}

\item{bootstrap}{If to use boostrap for VIP calculation, so that standard deviation on VIP can be estimated. Default is \code{TRUE}.}

\item{boot.n}{Set only when \code{boostrap = TRUE}, nummbers of iterations for boostrap. Default is \code{50}.}

\item{boot.parallelComputing}{Set only when \code{boostrap = TRUE}, if to use parallel computering for bootstrap process. Default is \code{TRUE}.}

\item{boot.n_cores}{Only set when \code{parallelComputing = TRUE}, the number of CPU cores to use. Default is \code{detectCores() - 1}, or the total number cores minus one.}

\item{boot.clusterType}{Set only when \code{boostrap = TRUE} and \code{boot.parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}

\item{plot}{If to generate a plot. Default is \code{TRUE}.}

\item{...}{Additional arguments to \code{\link{rbioFS_plsda_vip_plot}}.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}
}
\value{
Outputs a \code{rbiomvr_vip} class object to the environment.

\code{rbiomvr_vip} items are:

\code{vip_summary}

\code{comps}: number of PLS-DA components calculated fro VIP

\code{vip.alpha}: cutoff VIP value for a feature to be considered as important

\code{features_above_alpha}: selected features according to vip.alpha, i.e. features with VIP >= vip.alpha

\code{boostrap}

\code{boot.n}: boostrap iterations

\code{bootstrap.iteration.results}: raw VIP values for each bootstrap iteration
}
\description{
VIP, or variable importance in projection, calcualtion and plotting for plsda models. This is another FS method, and can be used independently.
}
\details{
Only works when the plsda model is fitted with the orthorgonal score algorithm, or NIPALS. Such model can be built using \code{\link{rbioClass_plsda}} with \code{method = "oscorespls"}.

The \code{vip.alpha} of 1 is the most commonly accepted value. However it is also acceptable to set according to the data and objectives of the study.
}
\examples{
\dontrun{
rbioFS_plsda_vip(object = new_model_optm, boostrap = TRUE, boot.m = 50,
                 vip.alpha = 0.8,
                 plot = TRUE, plot.title = TRUE, plot.titleSize = 10,
                 plot.sig.line = TRUE,
                 plot.outlineCol = "black", plot.errorbar = "SEM", plot.errorbarWidth = 0.2,
                 plot.errorbarLblSize = 6, plot.fontType = "sans", plot.xLabel = "Features",
                 plot.xLabelSize = 10, plot.xTickLblSize = 10, plot.xTickItalic = FALSE,
                 plot.xTickBold = FALSE, plot.xAngle = 90, plot.xhAlign = 1, plot.xvAligh = 0.2,
                 plot.rightsideY = TRUE,
                 plot.yLabel = "Coefficients", plot.yLabelSize = 10, plot.yTickLblSize = 10,
                 plot.yTickItalic = FALSE, plot.yTickBold = FALSE, plot.legendSize = 9,
                 plot.legendTtl = FALSE, plot.legendTtlSize = 9, plot.Width = 170,
                 plot.Height = 150)
}
}
