% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioFS_rf_sfs.R
\name{rbioFS_rf_SFS}
\alias{rbioFS_rf_SFS}
\title{rbioFS_rf_SFS}
\usage{
rbioFS_rf_SFS(objTitle = "x_vs_tgt", x, targetVar, nTimes = 50,
  nTree = 1001, mTry = "recur_default", multicore = TRUE,
  plot = TRUE, n = "all", ...)
}
\arguments{
\item{objTitle}{The title for the output data frame. Default is \code{"x_vs_tgt"}}

\item{x}{Input dataframe or matrix. Make sure to arrange the data with features as column names. Note this excludes the label column.}

\item{targetVar}{The target variable for random forest feature selection. This is a factor object.}

\item{nTimes}{Number of random forest vi computation runs. Default is \code{50} times.}

\item{nTree}{Number of trees generated for each random forest run. Default is \code{1001} trees.}

\item{mTry}{Number of randomly selected featurs for constructing trees. When \code{"recur_default"}, it'll be based on \code{p / 3}; when \code{"rf_default"}, it will use the default setting in \code{randomForest} package. Default is \code{"recur_default"}.}

\item{multicore}{If to use parallel computing. Default is \code{TRUE}.}

\item{plot}{If to plot a bargraph to visualize vi and the ranking. Default is \code{TRUE}}

\item{n}{Number of features to show. Takes integer numbers. Default is \code{"all"} (make sure to include quotation marks).}

\item{...}{Additional arguments passed to the plot function \code{\link{rbioFS_rf_SFS_plot}}.}
}
\value{
Outputs a \code{rf_sfs} object with  OOB error rate summary, and a joint-point curve in \code{csv} format.
}
\description{
Recursive nested random froest variable importance (vi) and OOB error rate computation in a sequential forward selection (SFS) manner.
}
\details{
Make sure to arrange data (dfm) with feature (e.g., gene) as variables (i.e., columns), and rownames as sample names.
}
\examples{
\dontrun{
rbioFS_rf_SFS(training_HCvTC, tgtVar_HCvTC, multicore = TRUE)
}
}
