% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioFS_svm.R
\name{rbioFS_svm}
\alias{rbioFS_svm}
\title{rbioFS_svm}
\usage{
rbioFS_svm(x, y, center = TRUE, scale = TRUE, kernel = "radial",
  tune.method = "cross", tune.cross.k = 10, tune.boot.n = 10, ...,
  verbose = TRUE)
}
\arguments{
\item{x}{Input data matrix (e.g., independent variables, predictors, features, X, etc). Make sure it is either a matrix or a dataframe.}

\item{y}{Input response variable (e.g.,dependent variables, Y etc). Make sure it is \code{factor} class.}

\item{center}{Logical, wether center the data, i.e. subtracting mean. Default is \code{TRUE}.}

\item{scale}{Logical, whether to scale the data, i.e. dividing by standard deviation. Default is \code{TRUE}.}

\item{kernel}{SVM kernel. Options are \code{"linear", "ploynomial", "radial", "sigmoid"}. Default is \code{"radial"}, aka RBF.}

\item{tune.method}{Parameter tuning method. Options are \code{"cross"} (i.e. cross validation), \code{"boot"} (i.e. bootstrap), and \code{"fix"}. Default is \code{"cross"}.}

\item{tune.cross.k}{Set only when \code{tune.method = "cross"}, fold number for cross validation. Default is \code{10}.}

\item{tune.boot.n}{Set only when \code{tune.method = "boot"}, bootstrap iterations. Default is \code{10}.}

\item{...}{Additional arguments for \code{svm} function from \code{e1071} pacakge.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will be affect error or warning messeages.}
}
\value{
Returns a SVM model object, with classes "svm" and "rbiosvm".
}
\description{
Support Vector Machine (SVM) modelling
}
\details{
Parameter tuning is for gamma (not applicable when \code{kernel = "linear"}) and cost.
}
\examples{
\dontrun{
svm_model <- rbioFS_svm(x = training_set[, -1], y = training_set[, 1], kernel = "radial", center = TRUE, scale = FALSE)
}
}
