% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioUtil.R
\name{rbioUtil_perm_plot}
\alias{rbioUtil_perm_plot}
\alias{rbioUtil_perm_plot.rbiosvm_perm}
\alias{rbioUtil_perm_plot.rbiomvr_perm}
\alias{rbioUtil_perm_plot.default}
\title{Title rbioUtil_perm_plot}
\usage{
rbioUtil_perm_plot(perm_res, ...)

\method{rbioUtil_perm_plot}{rbiosvm_perm}(perm_res, plot.SymbolSize = 2,
  plot.Width = 170, plot.Height = 150, ..., verbose = TRUE)

\method{rbioUtil_perm_plot}{rbiomvr_perm}(perm_res, plot.SymbolSize = 2,
  plot.Width = 170, plot.Height = 150, ..., verbose = TRUE)

\method{rbioUtil_perm_plot}{default}(baseplt, plot.display.Title = TRUE,
  plot.titleSize = 10, plot.fontType = "sans",
  plot.xLabel = "Permutations", plot.xLabelSize = 10,
  plot.xTickLblSize = 10, plot.yLabel = "Accuracy",
  plot.yLabelSize = 10, plot.yTickLblSize = 10, plot.legendSize = 9,
  plot.rightsideY = TRUE)
}
\arguments{
\item{perm_res}{Permutation test results object. The object should be either \code{rbiosvm_perm}, or \code{rbiomvr_perm} classes.}

\item{...}{Additional argument for the plot settings, see details.}

\item{plot.SymbolSize}{Symbol size. Default is \code{2}.}

\item{plot.Width}{Scoreplot width. Default is \code{170}.}

\item{plot.Height}{Scoreplot height. Default is \code{150}.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}

\item{plot.display.Title}{If to show the name of the y class. Default is \code{TRUE}.}

\item{plot.titleSize}{The font size of the plot title. Default is \code{10}.}

\item{plot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is avaiable on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{plot.xLabel}{X-axis label. Type with quotation marks. Could be NULL. Default is \code{"Permutation"}.}

\item{plot.xLabelSize}{X-axis label size. Default is \code{10}.}

\item{plot.xTickLblSize}{X-axis tick label size. Default is \code{10}.}

\item{plot.yLabel}{Y-axis label. Type with quotation marks. Could be NULL.}

\item{plot.yLabelSize}{Y-axis label size. Default is \code{10}.}

\item{plot.yTickLblSize}{Y-axis tick label size. Default is \code{10}.}

\item{plot.legendSize}{Legend size. Default is \code{9}.}

\item{plot.rightsideY}{If to show the right side y-axis. Default is \code{FALSE}.}

\item{perm_res}{Permutation test results object. The object should be \code{rbiosvm_perm} class.}

\item{...}{Additional argument for the plot settings.}

\item{perm_res}{Permutation test results object. The object should be \code{rbiomvr_perm} class.}

\item{plot.SymbolSize}{Symbol size. Default is \code{2}.}

\item{plot.Width}{Scoreplot width. Default is \code{170}.}

\item{plot.Height}{Scoreplot height. Default is \code{150}.}

\item{...}{Additional argument for the plot settings.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}

\item{baseplot}{A base plot generated by the \code{\link{rbioUtil_perm_plot}} method.}
}
\value{
A pdf file containing a scatter plot for permutation results.

A pdf file containing a scatter plot for permutation results.

A pdf file containing a scatter plot for permutation results.

A baseplot object for permutation results.
}
\description{
Plotting function for permutation test results.
}
\examples{

\dontrun{
rbioUtil_perm_plot(perm_res = svm_model_perm)
}

}
